% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_wy.R
\name{adjp_wysd}
\alias{adjp_wysd}
\title{Westfall Young Step Down Function}
\usage{
adjp_wysd(
  rawp.mat,
  B,
  Sigma,
  t.df,
  two.tailed,
  cl = NULL,
  verbose = TRUE,
  updateProgress = NULL
)
}
\arguments{
\item{rawp.mat}{a matrix of raw p-values under H1. dimension: nrow = tnum, ncol = M}

\item{B}{the number of samples for which p-values under the alternative hypothesis
are compared with the distribution (matrix) of p-values under the complete null (this distribution
is obtained through drawing test values under H0 with a default of 10,000)}

\item{Sigma}{correlation matrix of null p-values}

\item{t.df}{degrees of freedom of null p-values}

\item{two.tailed}{one or two-tailed test}

\item{cl}{clusters for parallel computing}

\item{verbose}{verbose messaging}
}
\value{
a matrix of adjusted p-values
}
\description{
This adjustment function utilizes the comp_rawp_ss helper function to compare
each row of the matrix sample p-values under
alternative hypothesis to all the rows in the matrix of the p-values
under the complete null.
}
