% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_wy.R
\name{adjp_wyss}
\alias{adjp_wyss}
\title{Westfall-Young Single Step Adjustment Function}
\usage{
adjp_wyss(
  rawp.mat,
  B,
  Sigma,
  t.df,
  two.tailed,
  verbose = TRUE,
  updateProgress = NULL
)
}
\arguments{
\item{rawp.mat}{a matrix of raw p-values under H1. 
dimension: nrow = tnum, ncol = M}

\item{B}{numer of WY permutations}

\item{Sigma}{correlation matrix of null p-values}

\item{t.df}{degrees of freedom of null p-values}

\item{two.tailed}{one or two-tailed test}

\item{verbose}{whether to print out messaging}

\item{updateProgress}{function to update progress bar 
(only used for PUMP shiny app)}
}
\value{
a matrix of adjusted p-values
}
\description{
This adjustment function utilizes the comp_rawp_ss 
helper function to compare
each row of the matrix sample p-values under
alternative hypothesis to all the rows in the matrix of the p-values
under the complete null.
}
\keyword{internal}
