% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{adjust.allsamps.WYSD}
\alias{adjust.allsamps.WYSD}
\title{Adjust allsamps WYSD}
\usage{
adjust.allsamps.WYSD(snum, abs.Zs.H0, abs.Zs.H1, order.matrix, ncl)
}
\arguments{
\item{snum}{blah blah}

\item{abs.Zs.H0}{blah blah}

\item{abs.Zs.H1}{blah blah}

\item{order.matrix}{blah blah}

\item{ncl}{blah blah}
}
\value{
blah blah
}
\description{
Adjust allsamps WYSD
}
