% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{adjust.allsamps.WYSS}
\alias{adjust.allsamps.WYSS}
\title{Adjust Single Step WestFallYoung}
\usage{
adjust.allsamps.WYSS(snum, abs.Zs.H0, abs.Zs.H1)
}
\arguments{
\item{snum}{blah blah}

\item{abs.Zs.H0}{blah blah}

\item{abs.Zs.H1}{blah blah}
}
\value{
blah blah
}
\description{
Adjust Single Step WestFallYoung
}
