% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{as.data.frame.pumpresult}
\alias{as.data.frame.pumpresult}
\alias{pumpresult}
\title{Cast pumpresult result to data.frame}
\usage{
\method{as.data.frame}{pumpresult}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{the pumpresult object to convert.}

\item{row.names}{NULL or a character vector giving the row names for the data frame.}

\item{optional}{logical. If TRUE, setting row names and converting column names is optional.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Cast pumpresult result to data.frame
}
\seealso{
Other pumpresult: 
\code{\link{design}()},
\code{\link{dim.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{print.pumpresult}()},
\code{\link{pump_type}()},
\code{\link{search_path}()},
\code{\link{summary.pumpresult}()}
}
\concept{pumpresult}
