% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{calc.df}
\alias{calc.df}
\title{This function calculates the degree of freedom for all implemented designs}
\usage{
calc.df(design, J, K, nbar, numCovar.1, numCovar.2, numCovar.3)
}
\arguments{
\item{design}{RCT design (see list/naming convention)}

\item{J}{the number of schools}

\item{K}{the number of districts}

\item{nbar}{units per block}

\item{numCovar.1}{number of Level 1 baseline covariates (not including block dummies)}

\item{numCovar.2}{number of Level 2 baseline covariates}

\item{numCovar.3}{number of Level 3 baseline covariates}
}
\value{
the degree of freedom
}
\description{
This function calculates the degree of freedom for all implemented designs
}
