% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{calc.J}
\alias{calc.J}
\title{This function calculates needed J for all implemented designs}
\usage{
calc.J(design, MT, MDES, nbar, Tbar, R2.1, R2.2, ICC.2, omega.2)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MT}{multiplier}

\item{MDES}{scalar, or vector of length M; the MDES values for each outcome}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar; school intraclass correlation}

\item{omega.2}{scalar; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}
}
\value{
J, the number of schools
}
\description{
This function calculates needed J for all implemented designs
}
