% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_sample.R
\name{calc.K}
\alias{calc.K}
\title{Calculates K, the number of districts}
\usage{
calc.K(
  design,
  MT,
  MDES,
  J,
  nbar,
  Tbar,
  R2.1,
  R2.2,
  R2.3,
  ICC.2,
  ICC.3,
  omega.2,
  omega.3
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MT}{multiplier}

\item{J}{scalar; the number of schools}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{Tbar}{scalar; the proportion of samples that are assigned to the treatment}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by Level 1 covariates for each outcome}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by Level 2 covariates for each outcome}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by Level 3 covariates for each outcome}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{omega.2}{scalar; ratio of school effect size variability to random effects
variability}

\item{omega.3}{scalar; ratio of district effect size variability to random effects
variability}

\item{vector}{of length M; the MDES values for each outcome.
Can provide single MDES value which will be repeated for the M outcomes.}
}
\value{
K, the number of districts
}
\description{
Calculates K, the number of districts
}
