% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_sample.R
\name{calc.nbar}
\alias{calc.nbar}
\title{This function calculates needed nbar to achieve a given power (as represented by
a difference in t-statistic) for all implemented designs}
\usage{
calc.nbar(
  design,
  MT = 2.8,
  MDES,
  J,
  K = NULL,
  Tbar,
  R2.1,
  R2.2,
  ICC.2,
  omega.2,
  R2.3 = NULL,
  ICC.3 = NULL,
  omega.3 = NULL
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MT}{Number of approximate effect-size unit SEs (adjusted for degrees of
freedom issues) that the MDES needs to be to achieve desired power.  E.g.,
2.8 for normal theory.}

\item{MDES}{scalar, or vector of length M; the MDES values for each outcome}

\item{K}{scalar; the number of districts}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar; school intraclass correlation}

\item{omega.2}{scalar; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome. Defaults to 0.}

\item{ICC.3}{scalar; district intraclass correlation}

\item{omega.3}{scalar; ratio of variance of district-average impacts to
variance of district-level random intercepts. Default to 0 (no treatment
variation).}
}
\value{
nbar, the number of individuals needed, or NA if not possible given design
}
\description{
This function calculates needed nbar to achieve a given power (as represented by
a difference in t-statistic) for all implemented designs
}
