% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{calc.Q.m}
\alias{calc.Q.m}
\title{Computes Q_m, the standard error of the effect size estimate}
\usage{
calc.Q.m(
  design,
  J,
  K,
  nbar,
  Tbar,
  R2.1,
  R2.2,
  R2.3,
  ICC.2,
  ICC.3,
  omega.2,
  omega.3
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{R2.1}{vector of length M; percent of variation explained by Level 1
covariates for each outcome}

\item{R2.2}{vector of length M; percent of variation explained by Level 2
covariates for each outcome}

\item{R2.3}{vector of length M; percent of variation explained by Level 3
covariates for each outcome}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{omega.2}{scalar; ratio of school effect size variability to random
effects variability}

\item{omega.3}{scalar; ratio of district effect size variability to random
effects variability}
}
\value{
Q_m, the standard error of the effect size estimate
}
\description{
Function to calculate the theoretical true standard error of a given design
and model, in effect size units.
}
