% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{calc_df}
\alias{calc_df}
\title{Calculate the degrees of freedom for a particular d_m}
\usage{
calc_df(d_m, J, K, nbar, numCovar.1, numCovar.2, numCovar.3, validate = TRUE)
}
\arguments{
\item{d_m}{string; a single design and model code}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{validate}{whether or not to validate if output df is <= 0}
}
\value{
Degree of freedom for the d_m.
}
\description{
Given sample sizes, return the used degrees of freedom (frequently
conservative) for the d_m.
}
