% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_sample.R
\name{calc_MT}
\alias{calc_MT}
\title{Caculate multiplier for MDE calculation}
\usage{
calc_MT(df, alpha, two.tailed, target.power)
}
\description{
Caculate multiplier for MDE calculation
}
\keyword{internal}
