% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{calc_pval}
\alias{calc_pval}
\title{Calculates p-values from t-values}
\usage{
calc_pval(rawt, t.df, two.tailed)
}
\arguments{
\item{rawt}{vector of t statistics}

\item{t.df}{degrees of freedom}

\item{two.tailed}{whether to calculate 1 or 2-tailed p-values}
}
\value{
power results for individual, minimum, complete power
}
\description{
Calculates p-values from t-values
}
\keyword{internal}
