% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{check_cor}
\alias{check_cor}
\title{Check correlation of test statistics (simulation function)}
\usage{
check_cor(d_m, model.params.list, Tbar, n.sims = 100)
}
\arguments{
\item{d_m}{string; a single context, which is a design and model code. 
See pump_info() for list of choices.}

\item{model.params.list}{list; model parameters such as ICC,
R2, etc. See simulation vignette for details.}

\item{Tbar}{scalar; the proportion of samples 
that are assigned to the treatment.}

\item{n.sims}{Number of simulated datasets to generate.
More datasets will achieve a more accurate result
but also increase computation time.}
}
\value{
dataframe; contains estimated pairwise
correlations between all outcomes, plus information
about input parameters.
}
\description{
Given a design, model, and set of model parameters,
estimates the pairwise correlations between test statistics for
all outcomes.
}
