% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{check_pump_call}
\alias{check_pump_call}
\title{Check user inputs}
\usage{
check_pump_call(
  design,
  call = c("power", "mdes", "sample"),
  verbose = TRUE,
  multi.MTP.ok = TRUE,
  ...
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{call}{String denoting intended pump_power, pump_mdes, or pump_sample
call.}

\item{verbose}{Keep and return smaller messages or not}

\item{multi.MTP.ok}{TRUE/FALSE on whether multiple MTP is allowed for check.}

\item{...}{The arguments to be passed to given call.}
}
\value{
list of two things: "ok", a TRUE/FALSE of whether this is a valid
  call.  "messages", a list of all messages, warnings, and notes that would
  be generated by the input parameter choices.  In the case of an error, the
  messages will have the error as the first string in the list.
}
\description{
This functions takes in a list of user inputs and checks them for validity,
producing a mix of errors or warnings.
}
