% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{comp.rawt.SD}
\alias{comp.rawt.SD}
\title{Helper Functions for WestFallYoung Step down}
\usage{
comp.rawt.SD(abs.Zs.H0.1row, abs.Zs.H1.1samp, oo)
}
\arguments{
\item{abs.Zs.H0.1row}{A vector of permutated test statistics values under H0}

\item{abs.Zs.H1.1samp}{One sample of raw statistics}

\item{oo}{Order matrix of test statistics in descending order}
}
\value{
returns a vector of 1s and 0s with lengths of M outcomes
}
\description{
Helper Functions for WestFallYoung Step down
}
