% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_wy.R
\name{comp_rawp_sd}
\alias{comp_rawp_sd}
\title{Helper function for Westfall Young Step Down}
\usage{
comp_rawp_sd(nullp, rawp, rawp.order)
}
\arguments{
\item{nullp}{a vector of p-values under H0}

\item{rawp}{a vector of p-values under H1}

\item{rawp.order}{order vector of raw p-values in ascending order}
}
\value{
returns a vector of 1s and 0s with lengths of M outcomes
}
\description{
Helper function for Westfall Young Step Down
}
\keyword{internal}
