% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{convert_params}
\alias{convert_params}
\title{Converts user-inputted parameters into relevant DGP parameters.}
\usage{
convert_params(model.params.list)
}
\arguments{
\item{model.params.list}{list of DGP parameters}
}
\value{
dgp.params.list
}
\description{
Converts user-inputted parameters into relevant DGP parameters.
}
