% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{design}
\alias{design}
\title{Get design for pump result}
\usage{
design(x, ...)
}
\value{
design used
}
\description{
Get design for pump result
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{dim.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{power_curve}()},
\code{\link{print.pumpresult}()},
\code{\link{search_path}()}
}
\concept{pumpresult}
