% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{df}
\alias{df}
\title{Degrees of Freedom}
\usage{
df(J, n.j, numCovar.1)
}
\arguments{
\item{J}{blah blah}

\item{n.j}{blah blah}

\item{numCovar.1}{blah}
}
\value{
blah blah
}
\description{
Degrees of Freedom
}
