% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{dim.pumpresult}
\alias{dim.pumpresult}
\title{Dimension of pumpresult}
\usage{
\method{dim}{pumpresult}(x, ...)
}
\arguments{
\item{x}{A pumpresult object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Dimension of pumpresult
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{design}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{print.pumpresult}()},
\code{\link{pump_type}()},
\code{\link{search_path}()},
\code{\link{summary.pumpresult}()}
}
\concept{pumpresult}
