% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_power.R
\name{estimate_power_curve}
\alias{estimate_power_curve}
\title{Calculate a power curve for sample size or mdes.}
\usage{
estimate_power_curve(
  p,
  low = NULL,
  high = NULL,
  high.max = 1000,
  grid.size = 5,
  tnum = 2000
)
}
\arguments{
\item{p}{pumpresult object}

\item{low}{Low end of grid}

\item{high}{High end of grid}

\item{high.max}{If no high provided, maximum possible high}

\item{grid.size}{Number of points in grid}

\item{tnum}{the number of test statistics (samples)}
}
\value{
List of powers for grid of points.
}
\description{
For a grid of points based on a passed sample or mdes pumpresult, estimate
power.
}
\keyword{internal}
