% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_power.R
\name{find_best}
\alias{find_best}
\title{Extract roots from quadratic curve based on given evaluated points}
\usage{
find_best(test.pts, target.power, gamma = 1.5)
}
\arguments{
\item{test.pts}{power evaluated at different points}

\item{target.power}{goal power}
}
\value{
List of estimate of when curve reaches target.power, derivative of
  curve at that point, and parameters of the fit curve.
}
\description{
Extract roots from quadratic curve based on given evaluated points
}
