% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_power.R
\name{fit_bounded_logistic}
\alias{fit_bounded_logistic}
\title{Fit a bounded logistic curve}
\usage{
fit_bounded_logistic(x, y, wt)
}
\value{
Vector of four estimated parameters for the logistic curve: beta0, beta1, pmin, pmax
}
\description{
Curve is of form f(y) = pmin + (pmax-pmin) * logistic( beta0 + beta1*x )
}
\details{
(logistic as defined by plogis)
}
