% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_assignments}
\alias{gen_assignments}
\title{Generates school and district assignments (simulation function)}
\usage{
gen_assignments(J, K, nbar)
}
\arguments{
\item{J}{number of schools per district}

\item{K}{number of districts}

\item{nbar}{number of individuals per school}
}
\value{
list(S.id, D.id) of school and district 
assignments for each individual
}
\description{
Generates simple default schools and 
districts IDs for individual students for the purpose of
simulations. This assumes equal sized schools in 
equal sized districts.

This function is beyond the main scope of calculating power,
and is instead used for simulating data.
For more info on use, see the simulation vignette.
}
