% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_cluster_ids}
\alias{gen_cluster_ids}
\title{Generates school and district assignments (simulation
  function)}
\usage{
gen_cluster_ids(nbar, J, K)
}
\arguments{
\item{nbar}{scalar; number of individuals per school.}

\item{J}{scalar; number of schools per district.}

\item{K}{scalar; number of districts.}
}
\value{
list; school and district assignments (S.id, D.id) for each
  individual.
}
\description{
Generates simple default schools and districts IDs for
  individual students for the purpose of simulations. This assumes
  equal sized schools in equal sized districts.

  This function is beyond the main scope of calculating power, and
  is instead used for simulating data. For more info on use, see
  the simulation vignette.
}
