% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_corr_matrix}
\alias{gen_corr_matrix}
\title{Generate correlation matrix (simulation function)}
\usage{
gen_corr_matrix(M, rho.scalar)
}
\arguments{
\item{M}{dimension of matrix}

\item{rho.scalar}{fixed rho value}
}
\value{
rho.matrix, M x M matrix with rho.scalar as diag
}
\description{
Generate correlation matrix (simulation function)
}
