% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_cov_matrix}
\alias{gen_cov_matrix}
\title{generate covariance matrix between two variables}
\usage{
gen_cov_matrix(D, var1.vec, var2.vec, rho.matrix)
}
\arguments{
\item{D}{dimension of matrix}

\item{var1.vec}{vector of variances of first variable}

\item{var2.vec}{vector of variances of second variable}

\item{rho.matrix}{matrix of correlations}
}
\value{
Sigma matrix of covariance
}
\description{
generate covariance matrix between two variables
}
\keyword{internal}
