% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_full_data}
\alias{gen_full_data}
\title{Generate simulated multi-level data:
both unobserved and observed potential outcomes}
\usage{
gen_full_data(dgp.params.list)
}
\arguments{
\item{dgp.params.list}{list of data generating parameters}
}
\value{
list of: potential outcomes given control y0, treatment y1,
        covariates V.k, X.jk, C.ijk
        
        Each block has 1 column for each outcome (called "domain")
}
\description{
Generate simulated multi-level data:
both unobserved and observed potential outcomes
}
