% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_RE_cov_matrix}
\alias{gen_RE_cov_matrix}
\title{generate a parameterized covariance matrix from the provided 3 blocks}
\usage{
gen_RE_cov_matrix(Sigma.w, Sigma.z, Sigma.wz)
}
\arguments{
\item{Sigma.w}{level 3 covariance matrix}

\item{Sigma.z}{level 2 covariance matrix}

\item{Sigma.wz}{covariance between level 2 and level 3}
}
\value{
2M x 2M matrix for generating correlated pairs of random effects
}
\description{
generate a parameterized covariance matrix from the provided 3 blocks
}
\keyword{internal}
