% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_sim_data}
\alias{gen_sim_data}
\title{Generate simulated multi-level data (simulation function)}
\usage{
gen_sim_data(d_m, model.params.list, Tbar = 0.5)
}
\arguments{
\item{d_m}{string; a single context, which is a design and model code. 
See pump_info() for list of choices.}

\item{model.params.list}{list; model parameters such as ICC,
R2, etc. See simulation vignette for details.}

\item{Tbar}{scalar; the proportion of samples 
that are assigned to the treatment.}
}
\value{
list; potential outcomes, covariates,
observed outcomes, and treatment assignment.
}
\description{
Generates simulated data for multi-level
RCTs for pump-suppored designs and models for
both unobserved and observed potential outcomes.

Takes in a list of necessary data-generating parameters.

This function is beyond the main scope of calculating power,
and is instead used for simulating data.
For more info on use, see the simulation vignette.
}
