% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_T.x}
\alias{gen_T.x}
\title{Generate treatment assignment vector (simulation function)}
\usage{
gen_T.x(d_m, S.id, D.id, nbar, Tbar)
}
\arguments{
\item{d_m}{design and model}

\item{S.id}{vector of school assignments}

\item{D.id}{vector of district assignments}

\item{nbar}{number of level 1 units}

\item{Tbar}{probability of treatment assignment}
}
\value{
vector of treatment assignments for each unit
}
\description{
Given a RCT design and supporting information,
generates treatment assignments for each student.

This function is beyond the main scope of calculating power,
and is instead used for simulating data.
For more info on use, see the simulation vignette.
}
