% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_Yobs}
\alias{gen_Yobs}
\title{Generate observed outcomes (simulations)}
\usage{
gen_Yobs(full.data, T.x)
}
\arguments{
\item{full.data}{full dataset of potential outcomes}

\item{T.x}{N-vector of binary assignment to treat/control}
}
\value{
Yobs
}
\description{
Takes in a full dataset of both observed
and latent potential outcomes and the treatment assignment vector,
and returns only the observed outcomes.
}
