% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_wy.R
\name{get.adjp.minp}
\alias{get.adjp.minp}
\title{Helper function for Westfall Young}
\usage{
get.adjp.minp(ind.B, rawt.order)
}
\arguments{
\item{ind.B}{matrix of indicator variables for if each raw test statistic exceeds
the null test statistics. dimensions: nrow = tnum, ncol = M.}

\item{rawt.order}{order of raw test statistics in descending order}
}
\value{
returns adjusted p-value matrix
}
\description{
enforces monotonicity in p-values.
}
