% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{get.power.results}
\alias{get.power.results}
\title{Calculates different definitions of power}
\usage{
get.power.results(pval.mat, ind.nonzero, alpha)
}
\arguments{
\item{pval.mat}{matrix of p-values, columns are outcomes}

\item{ind.nonzero}{vector indicating which outcomes are nonzero}

\item{alpha}{scalar; the family wise error rate (FWER)}
}
\value{
power results for individual, minimum, complete power
}
\description{
This function takes in a matrix of adjusted p-values and outputs different types of power
}
