% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_wy.R
\name{get_adjp_minp}
\alias{get_adjp_minp}
\title{Helper function for Westfall Young}
\usage{
get_adjp_minp(ind.B, rawp.order)
}
\arguments{
\item{ind.B}{matrix of indicator variables for 
if each raw test statistic exceeds
the null test statistics. 
dimensions: nrow = tnum, ncol = M.}

\item{rawp.order}{order of raw p-values in ascending order}
}
\value{
returns adjusted p-value matrix
}
\description{
enforces monotonicity in p-values.
}
\keyword{internal}
