% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_pval_tstat}
\alias{get_pval_tstat}
\title{Function: get_pval_tstat	                                       
 
extracts p-value and t statistics from a given model}
\usage{
get_pval_tstat(mod, d_m, model.params.list)
}
\arguments{
\item{mod}{model object}

\item{model.params.list}{list of model parameters}
}
\description{
Function: get_pval_tstat	                                       
 
extracts p-value and t statistics from a given model
}
\keyword{internal}
