% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_rawpt}
\alias{get_rawpt}
\title{Function: get_rawpt                                      
 
fits models and extracts p values and t statistics}
\usage{
get_rawpt(dat.all, d_m, model.params.list)
}
\arguments{
\item{dat.all}{list of datasets (of length M)}

\item{d_m}{design/model}

\item{model.params.list}{list of model parameters}
}
\description{
Function: get_rawpt                                      
 
fits models and extracts p values and t statistics
}
\keyword{internal}
