% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{is.pumpresult}
\alias{is.pumpresult}
\alias{pumpresult}
\title{Is object a pumpresult object?}
\usage{
is.pumpresult(x)
}
\arguments{
\item{x}{the object to check.}
}
\description{
Is object a pumpresult object?
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{dim.pumpresult}()},
\code{\link{params}()},
\code{\link{print.pumpresult}()},
\code{\link{search_path}()}
}
\concept{pumpresult}
