% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{make_model}
\alias{make_model}
\title{Function: make.model		                                       
 
 function to generate a model for simulated data}
\usage{
make_model(dat.m, d_m)
}
\arguments{
\item{dat.m}{a data frame for a single outcome}

\item{d_m}{design/model}
}
\description{
Function: make.model		                                       
 
 function to generate a model for simulated data
}
\keyword{internal}
