% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{makelist_samp}
\alias{makelist_samp}
\title{takes in multi-outcome data and returns a list for each outcome}
\usage{
makelist_samp(samp.obs, T.x)

makelist_samp(samp.obs, T.x)
}
\arguments{
\item{samp.obs}{a list of observed data components}

\item{T.x}{vector of treatment assignments}
}
\description{
takes in multi-outcome data and returns a list for each outcome

takes in multi-outcome data and returns a list for each outcome
}
\keyword{internal}
