% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{MDES.blockedRCT.2}
\alias{MDES.blockedRCT.2}
\title{MDES function}
\usage{
MDES.blockedRCT.2(M, numFalse, Ai_mdes, J, n.j, power, power.definition,
  MTP, marginError, p, alpha, numCovar.1, numCovar.2 = 0, R2.1, R2.2,
  ICC, mod.type, sigma, omega, tnum = 10000, snum = 2, ncl = 2,
  updateProgress = NULL)
}
\arguments{
\item{M}{blah blah}

\item{numFalse}{the number of false numbers. Kristin has calculated it one way, but, I have structured it another using AiImpacts.}

\item{Ai_mdes}{blah blah}

\item{J}{blah blah}

\item{n.j}{blah blah}

\item{power}{blah blah}

\item{power.definition}{blah blah}

\item{MTP}{blah blah}

\item{marginError}{blah blah}

\item{p}{blah blah}

\item{alpha}{blah blah}

\item{numCovar.1}{blah blah}

\item{numCovar.2}{blah blah}

\item{R2.1}{blah blah}

\item{R2.2}{blah blah}

\item{ICC}{blah blah}

\item{mod.type}{blah blah}

\item{sigma}{blah blah}

\item{omega}{blah blah}

\item{tnum}{blah blah}

\item{snum}{blah blah}

\item{ncl}{blah blah}

\item{updateProgress}{this is the progress bar function that will be passed to the main MDES calculation function}
}
\value{
mdes results
}
\description{
MDES function
}
