% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_i1_2cfr.R
\name{mdes_blocked_i1_2c}
\alias{mdes_blocked_i1_2c}
\title{MDES (minimum detectable effect size) function}
\usage{
mdes_blocked_i1_2c(M, numFalse, J, n.j, power, power.definition, MTP,
  marginError, p, alpha, numCovar.1, numCovar.2 = 0, R2.1, R2.2, ICC,
  mod.type, sigma = 0, rho = 0.99, omega, tnum = 10000,
  snum = 1000, ncl = 2, updateProgress = NULL)
}
\arguments{
\item{M}{the number of hypothesis tests (outcomes)}

\item{numFalse}{the number of false nulls. This parameter is used for non-Shiny calculations.}

\item{J}{the number of blocks}

\item{n.j}{the harmonic mean of the number of units per block}

\item{power}{required statistical power for the experiment}

\item{power.definition}{definition of statistical power from individual, d-minimal to complete power}

\item{MTP}{type of multiple testing procedure}

\item{marginError}{the margin of error for MDES estimation based on targeted power value}

\item{p}{the proportion of samples that are assigned to the treatment}

\item{alpha}{the family wise error rate (FWER)}

\item{numCovar.1}{number of Level 1 baseline covariates (not including block dummies)}

\item{numCovar.2}{number of Level 2 baseline covariates (set to 0 for this design)}

\item{R2.1}{a vector of length M corresponding to R^2 for M outcomes of Level 1 (R^2 = variation in the data explained by the model)}

\item{R2.2}{a vector of length M corresponding to R^2 for M outcomes of Level 2 (R^2 = variation in the data explained by the model)}

\item{ICC}{intraclass correlation}

\item{mod.type}{"c" for constant effects, "f" for fixed effects, "r" for random effects (parameter not in use at the moment)}

\item{sigma}{correlation matrix for correlations between test statistics (parameter not in use at the moment.Default is set to 0.99)}

\item{rho}{correlation between outcomes. This generates the sigma matrix.}

\item{omega}{NULL (parameter in development)}

\item{tnum}{the number of test statistics (samples) for all procedures other than Westfall-Young & number of permutations for WY. The default is set at 10,000.}

\item{snum}{the number of samples for Westfall-Young. The default is set at 1,000.}

\item{ncl}{ncl the number of clusters to use for parallel processing. The default is set at 2.}

\item{updateProgress}{this is the progress bar function that will be passed to the main MDES calculation function}
}
\value{
mdes results
}
\description{
The minimum detectable effect size function calculates the most feasible minimum detectable effect size
for a given MTP, power and power definition. The goal is to find the MDES value that satisfies the margin of error
set in the parameter in the power value.
}
