% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{midpoint}
\alias{midpoint}
\title{Midpoint function}
\usage{
midpoint(lower, upper)
}
\arguments{
\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
returns midpoint value
}
\description{
Calculating the midpoint between the lower and upper bound by calculating half the distance between the two
and adding the lower bound to it. The function is a helper function in determining the MDES that falls within
acceptable power range.
}
