% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{optimize_power}
\alias{optimize_power}
\title{Optimizes power to help in search for MDES or SS}
\usage{
optimize_power(
  design,
  search.type,
  MTP,
  target.power,
  power.definition,
  tol,
  start.tnum,
  start.low,
  start.high,
  MDES = NULL,
  J = NULL,
  K = NULL,
  nbar = NULL,
  M = M,
  Tbar = Tbar,
  alpha,
  numCovar.1,
  numCovar.2,
  numCovar.3 = NULL,
  R2.1,
  R2.2,
  R2.3 = NULL,
  ICC.2,
  ICC.3 = NULL,
  omega.2,
  omega.3 = NULL,
  rho,
  B = NULL,
  cl = NULL,
  max.steps = 20,
  max.cum.tnum = 5000,
  final.tnum = 10000
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{search.type}{options: MDES, J, K (nbar not currently supported)}

\item{MTP}{a single multiple adjustment procedure of interest. Supported options: Bonferroni, BH,
Holm, WY-SS, WY-SD}

\item{target.power}{Target power to arrive at}

\item{power.definition}{must be a valid power type outputted by power function, i.e. D1indiv, min1, etc.}

\item{tol}{tolerance for target power}

\item{start.tnum}{number of samples for initial power search (should be low to increase speed)}

\item{start.low}{lower bound for possible power}

\item{MDES}{scalar, or vector of length M; the MDES values for each outcome.}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{Tbar}{scalar; the proportion of samples that are assigned to the treatment}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not including block
dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by Level 1 covariates for each outcome}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by Level 2 covariates for each outcome}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by Level 3 covariates for each outcome}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{omega.2}{scalar; ratio of school effect size variability to random effects variability}

\item{omega.3}{scalar; ratio of district effect size variability to random effects variability}

\item{rho}{scalar; correlation between outcomes}

\item{B}{scalar; the number of samples/permutations for Westfall-Young}

\item{cl}{cluster object to use for parallel processing}

\item{max.steps}{how many steps allowed before terminating}

\item{max.cum.tnum}{maximum cumulative number of samples}

\item{final.tnum}{number of samples for final draw}

\item{start.hight}{upper bound for possible power}

\item{tnum}{scalar; the number of test statistics (samples)}
}
\value{
power
}
\description{
Optimizes power to help in search for MDES or SS
}
