% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{parse_d_m}
\alias{parse_d_m}
\title{Return characteristics of a given d_m code}
\usage{
parse_d_m(d_m)
}
\arguments{
\item{d_m}{String. Experimental d_m to parse.}
}
\value{
List of features including number of levels, level of randomization,
  etc.
}
\description{
See the pump_info method to get a list of supported d_ms.
}
\examples{
supported = pump_info()$d_m
supported$Code[[4]]
parse_d_m( supported$Code[[4]] )

}
\concept{pump_info}
