% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{parse_design}
\alias{parse_design}
\title{Return characteristics of a given design code}
\usage{
parse_design(design)
}
\arguments{
\item{design}{String. Experimental design to parse.}
}
\value{
List of features including number of levels, level of randomization,
  etc.
}
\description{
Return characteristics of a given design code
}
\concept{pump_info}
