% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parse_power_definition}
\alias{parse_power_definition}
\title{Parse the power definition}
\usage{
parse_power_definition(power.definition, M)
}
\arguments{
\item{power.definition}{i.e. D1indiv, min1, complete}

\item{M}{number of outcomes}
}
\value{
information about power type
}
\description{
Parse the power definition
}
\keyword{internal}
