% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pumpgrid}
\alias{plot.pumpgrid}
\title{Plot a pump grid object}
\usage{
\method{plot}{pumpgrid}(x, power.definition, var.vary)
}
\arguments{
\item{x}{pumpgrid object}
}
\description{
Plot a pump grid object
}
