% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pumpgridresult.mdes}
\alias{plot.pumpgridresult.mdes}
\title{Plot a grid pump mdes object}
\usage{
\method{plot}{pumpgridresult.mdes}(x, power.definition, var.vary, ...)
}
\arguments{
\item{x}{pumpgridresult object}

\item{power.definition}{definition of power to plot.  If NULL, plot all
definitions as a facet wrap.}

\item{var.vary}{variable to vary on X axis}

\item{...}{additional parameters}
}
\description{
Plot a grid pump mdes object
}
\keyword{internal}
