% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pumpgridresult}
\alias{plot.pumpgridresult}
\title{Plot a pump grid result object}
\usage{
\method{plot}{pumpgridresult}(x, power.definition, var.vary, ...)
}
\arguments{
\item{x}{pumpgridresult object}
}
\description{
Plot a pump grid result object
}
