% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pumpresult}
\alias{plot.pumpresult}
\title{Plot a single scenario pump object (result function)}
\usage{
\method{plot}{pumpresult}(x, type = "power", ...)
}
\arguments{
\item{x}{pumpresult object.}

\item{type}{string; "power" or "search".
Specifies whether to plot the default power graph,
or the search path. The search path is only valid
for MDES and SS results.}

\item{...}{additional parameters.}
}
\value{
plot; a ggplot object of power across
differen definitions.
}
\description{
Works on an object returned by pump_power(),
and visualizes different definitions of power across
MTPs. This function does not apply to pump_mdes()
or pump_sample() objects, as these functions only
return a single value.
}
\examples{
pp1 <- pump_power(d_m = "d2.2_m2rc", MTP = 'HO',
 nbar = 50, J = 20, M = 8, numZero = 5,
 MDES = 0.30, Tbar = 0.5, alpha = 0.05, two.tailed = FALSE,
 numCovar.1 = 1, numCovar.2 = 1, R2.1 = 0.1, R2.2 = 0.7, 
 ICC.2 = 0.05, rho = 0.2, tnum = 5000)
 
plot(pp1)

mdes <- pump_mdes(d_m = "d2.1_m2fc", MTP = 'HO',
  power.definition = 'D1indiv', target.power = 0.7,
  J = 60, nbar = 50, M = 3, Tbar = 0.5, alpha = 0.05,
  numCovar.1 = 1, R2.1 = 0.1, ICC.2 = 0.05, rho = 0.2)
  
plot(mdes)

J <- pump_sample(d_m = "d2.1_m2fc",
   MTP = 'HO', power.definition = 'D1indiv',
   typesample = 'J', target.power = 0.6,
   nbar = 50, M = 3, MDES = 0.125,
   Tbar = 0.5, alpha = 0.05,
   numCovar.1 = 1, R2.1 = 0.1, ICC.2 = 0.05, 
   rho = 0.2, tnum = 1000)
plot(J, type = "search")

}
