% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_power.R
\name{plot_power_curve}
\alias{plot_power_curve}
\title{#' Extract roots from quadratic curve based on given evaluated points
#'
#' @param test.pts power evaluated at different points
#' @param start.low lower bound
#' @param start.high upper bound
#' @param target.power goal power
#' @param alternate alternate point to return if quadratic fit fails
#'
#' @return root of quadratic curve}
\usage{
plot_power_curve(pwr, plot.points = TRUE)
}
\arguments{
\item{pwr}{Result from the pump_sample or pump_mdes}

\item{plot.points}{flag; whether to plot individual points on curve}
}
\description{

}
\details{

}
