% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power_search}
\alias{plot_power_search}
\title{Examine search path of the power search.}
\usage{
plot_power_search(pwr)
}
\arguments{
\item{pwr}{Result from a pump_sample or pump_mdes call.}
}
\value{
a ggplot plot (a gridExtra arrangement of 3 plots, technically).
}
\description{
This will give a triple-plots about how the search narrowed down into the final
estimate.  Can be useful to gauge where convergence went poorly.
}
