% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{power.blockedRCT.2}
\alias{power.blockedRCT.2}
\title{Block RCT2 power function}
\usage{
power.blockedRCT.2(M, MDES, J, n.j, p, alpha, numCovar.1, numCovar.2 = 0,
  R2.1, R2.2, ICC, mod.type, sigma, omega, tnum = 10000, snum = 1000,
  ncl = 2)
}
\arguments{
\item{M}{the number of hypothesis tests (outcomes)}

\item{MDES}{vector of MDES's of length M - can be zero if not assuming all nulls are false}

\item{J}{number of blocks}

\item{n.j}{units per block}

\item{p}{the proportion of samples that are assigned the treatment}

\item{alpha}{significance level}

\item{numCovar.1}{number of Level 1 baseline covariates (not including block dummies)}

\item{numCovar.2}{number of Level 2 baseline covariates}

\item{R2.1}{a vector of length M corresponding to R^2 for M outcomes of Level 1(R^2 = variation in the data explained by the model)}

\item{R2.2}{a vector of length M corresponding to R^2 for M outcomes of Level 2}

\item{ICC}{intraclass correlation}

\item{mod.type}{"c" for constant effects, "f" for fixed effects, "r" for random effects}

\item{sigma}{correlation matrix for correlations between test statistics (this need to be flexible across multiple test statistics. Now, it's set to be at 1)}

\item{omega}{NULL}

\item{tnum}{number of test statistics (samples) for all procedures other than WY & number of permutations for WY (i.e permutation samples) WY has permutation which is reassignment of treatment variables. We do not see the treatment reassignment for permutation explicitly here.}

\item{snum}{number of samples for WY (i.e are they less than the other test. If so, why?)}

\item{ncl}{the number of clusters to use for parallel processing. It has a default of 2.}
}
\value{
power results
}
\description{
Block RCT2 power function
}
