% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{power_curve}
\alias{power_curve}
\title{Obtain power curve over a range of parameters}
\usage{
power_curve(
  x,
  all = FALSE,
  low = NULL,
  high = NULL,
  grid.size = 5,
  tnum = 2000
)
}
\arguments{
\item{x}{A pumpresult object.}

\item{all}{Merge in the search path from the original search.}

\item{low}{Low end of grid}

\item{high}{High end of grid}

\item{grid.size}{Number of points in grid}

\item{tnum}{the number of test statistics (samples)}
}
\description{
This is used to see rate of power change.
}
