% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{print.pumpresult}
\alias{print.pumpresult}
\title{Pretty print pump result}
\usage{
\method{print}{pumpresult}(x, n = 10, no_header = FALSE, ...)
}
\arguments{
\item{x}{A pumpresult object.}

\item{...}{No extra options passed.}
}
\description{
Pretty print pump result
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{design}()},
\code{\link{dim.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{pump_type}()},
\code{\link{search_path}()},
\code{\link{summary.pumpresult}()}
}
\concept{pumpresult}
