% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{print_context}
\alias{print_context}
\title{Print context of given pump result object}
\usage{
print_context(x, insert_results = FALSE, insert_control = FALSE, ...)
}
\arguments{
\item{x}{A pumpresult object.}

\item{insert_results}{Include actual results in the printout.}

\item{insert_control}{Include the optimizer control parameter information.}

\item{...}{Extra arguments to pass to print.pumpresult.}
}
\description{
Print context of given pump result object
}
