% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{print_design}
\alias{print_design}
\title{Print design of given pump result object}
\usage{
print_design(x, insert_results = FALSE, ...)
}
\arguments{
\item{x}{A pumpresult object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Print design of given pump result object
}
