% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\docType{package}
\name{PUMP}
\alias{PUMP}
\title{PUMP: A package for estimating power under multiplicity}
\description{
The PUMP package provides three core functions:
\itemize{
\item pump_power() for estimating power
\item pump_mdes() for estimating minimum detectable effect size
\item pump_sample() for estimating sample size.
}
}
\details{
For a full package description, see \url{https://arxiv.org/abs/2112.15273}.
}
