% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_mdes.R
\name{pump_mdes}
\alias{pump_mdes}
\title{MDES (minimum detectable effect size) function}
\usage{
pump_mdes(
  design,
  MTP,
  M,
  J,
  K = 1,
  target.power,
  power.definition,
  tol,
  nbar,
  Tbar,
  alpha,
  numCovar.1 = 0,
  numCovar.2 = 0,
  numCovar.3 = 0,
  R2.1 = 0,
  R2.2 = 0,
  R2.3 = 0,
  ICC.2 = 0,
  ICC.3 = 0,
  rho,
  omega.2 = 0,
  omega.3 = 0,
  tnum = 10000,
  B = 1000,
  max.steps = 20,
  max.cum.tnum = 5000,
  start.tnum = 200,
  final.tnum = 10000,
  cl = NULL,
  updateProgress = NULL
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MTP}{a single multiple adjustment procedure of interest. Supported
options: Bonferroni, BH, Holm, WY-SS, WY-SD}

\item{M}{scalar; the number of hypothesis tests (outcomes)}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{target.power}{Target power to arrive at}

\item{power.definition}{must be a valid power type outputted by power
function, i.e. D1indiv, min1, etc.}

\item{tol}{tolerance for target power}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{rho}{scalar; correlation between outcomes}

\item{omega.2}{scalar; ratio of school effect size variability to random
effects variability}

\item{omega.3}{scalar; ratio of district effect size variability to random
effects variability}

\item{tnum}{scalar; the number of test statistics (samples)}

\item{B}{scalar; the number of samples/permutations for Westfall-Young}

\item{max.steps}{how many steps allowed before terminating}

\item{max.cum.tnum}{maximum cumulative number of samples}

\item{final.tnum}{number of samples for final draw}

\item{cl}{cluster object to use for parallel processing}

\item{updateProgress}{the callback function to update the progress bar (User
does not have to input anything)}
}
\value{
mdes results
}
\description{
The minimum detectable effect size function calculates the most feasible
minimum detectable effect size for a given MTP, power and power definition.
The goal is to find the MDES value that satisfies the tolerance set in the
parameter in the power value.
}
