% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_sample.R
\name{pump_sample_raw}
\alias{pump_sample_raw}
\title{Calculating Sample for Raw (Unadjusted)}
\usage{
pump_sample_raw(
  design,
  MTP,
  typesample,
  MDES,
  M,
  nbar = NULL,
  J = NULL,
  K = NULL,
  target.power,
  Tbar,
  alpha,
  two.tailed = TRUE,
  numCovar.1 = 0,
  numCovar.2 = 0,
  numCovar.3 = 0,
  R2.1,
  R2.2 = NULL,
  R2.3 = NULL,
  ICC.2 = NULL,
  ICC.3 = NULL,
  rho = NULL,
  omega.2 = NULL,
  omega.3 = NULL,
  tol = 0.1,
  max.steps = 100
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MTP}{Multiple adjustment procedure of interest. Supported options:
Bonferroni, BH, Holm, WY-SS, WY-SD (passed as strings).  Provide list to
automatically re-run for each procedure on the list.}

\item{typesample}{type of sample size to calculate: J, K, or nbar}

\item{MDES}{scalar, or vector of length M; the MDES values for each outcome.}

\item{M}{scalar; the number of hypothesis tests (outcomes)}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{target.power}{target power to arrive at}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{two.tailed}{whether to calculate two-tailed or one-tailed power}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{rho}{scalar; assumed correlation between the test statistics of the
tests.}

\item{omega.2}{scalar; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}

\item{omega.3}{scalar; ratio of variance of district-average impacts to
variance of district-level random intercepts. Default to 0 (no treatment
variation).}

\item{tol}{tolerance of how much our search can change the sample size before
stopping.  (In number of units, so 0.1 is relatively precise.)}

\item{max.steps}{how many steps allowed before terminating}
}
\value{
raw sample returns
}
\description{
This is a Helper function for getting Sample Size when no adjustments has
been made to the test statistics.
}
\details{
It requires iteration because we do not know the degrees of freedom, and so
we guess and then calculate sample size, and then recalculate df based on
sample size, until we converge.
}
