% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_sample.R
\name{pump_sample_raw_old}
\alias{pump_sample_raw_old}
\title{stopping.  (In number of units, so 0.1 is relatively precise.)}
\usage{
pump_sample_raw_old(
  design,
  MTP,
  typesample,
  MDES,
  nbar = NULL,
  J = NULL,
  K = NULL,
  target.power,
  Tbar,
  alpha,
  two.tailed = TRUE,
  numCovar.1 = 0,
  numCovar.2 = 0,
  numCovar.3 = 0,
  R2.1,
  R2.2 = NULL,
  R2.3 = NULL,
  ICC.2 = NULL,
  ICC.3 = NULL,
  omega.2 = NULL,
  omega.3 = NULL,
  max.steps = 100
)
}
\description{
stopping.  (In number of units, so 0.1 is relatively precise.)
}
