% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{pump_type}
\alias{pump_type}
\title{What type of pumpresult}
\usage{
pump_type(x)
}
\value{
power, mdes, or sample, as a string.
}
\description{
What type of pumpresult
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{design}()},
\code{\link{dim.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{print.pumpresult}()},
\code{\link{search_path}()},
\code{\link{summary.pumpresult}()}
}
\concept{pumpresult}
