% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{pumpgrid}
\alias{pumpgrid}
\alias{is.pumpgrid}
\title{pumpresult object for results of grid power calculations}
\usage{
is.pumpgrid(x)
}
\arguments{
\item{x}{a pumpgrid object (except for is.pumpresult, where it is a generic
object to check).}
}
\value{
is.pumpresult: TRUE if object is a pumpgrid object.
}
\description{
The pumpgrid object is an S3 class that holds the results from
`pump_power_grid()`, `pump_sample_grid()`, and `pump_mdes_grid()`.

It has several methods that pull different information from this object, and
some printing methods for getting nicely formatted results.
}
