% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_functions.R
\name{run_grid}
\alias{run_grid}
\title{Run grid across any of the core pump functions}
\usage{
run_grid(
  args,
  pum_function,
  verbose = FALSE,
  drop_unique_columns,
  ...,
  use_furrr = FALSE
)
}
\arguments{
\item{args}{list of scenario arguments}

\item{pum_function}{pump_mdes, pump_sample, pump_power}

\item{verbose}{print out detailed diagnostics}

\item{drop_unique_columns}{TODO}

\item{...}{Extra arguments passed to the underlying pump_power, pump_sample,
or pump_mdes functions.}

\item{use_furrr}{not currently in use, whether to parallelize}
}
\description{
Run grid across any of the core pump functions
}
