% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_i1_2cfr.R
\name{sample_blocked_i1_2c}
\alias{sample_blocked_i1_2c}
\title{Sample Function}
\usage{
sample_blocked_i1_2c(M, numFalse, typesample, J, n.j, J0, n.j0, MDES,
  power, power.definition, MTP, marginError, p, alpha, numCovar.1,
  numCovar.2 = 0, R2.1, R2.2, ICC, mod.type, sigma = 0, rho = 0.99,
  omega, tnum = 10000, snum = 2, ncl = 2, num.iter = 20,
  updateProgress = NULL)
}
\arguments{
\item{M}{the number of hypothesis tests (outcomes)}

\item{numFalse}{the number of false nulls. This parameter is used for non-Shiny calculations}

\item{typesample}{the type of the number of sample we would like to estimate: either block J or n.j (harmonic mean within block. For Shiny use)}

\item{J}{the number of blocks (set to NULL if you do not want to estimate this one)}

\item{n.j}{the harmonic mean of blocks (set to NULL if you do not want to estimate this one)}

\item{J0}{the initial value for the sample number of blocks. The default is set at 10.}

\item{n.j0}{the initial value for the harmonic mean for the number of samples within block. The default is set at 10.}

\item{MDES}{minimum detectable effet size}

\item{power}{required statistical power for the experiment}

\item{power.definition}{definition of statistical power from individual, d-minimal to complete power}

\item{MTP}{type of multiple testing procedure in use from Bonferroni, Benjamini-Hocheberg, Holms, Westfall-Young Single Step, Westfall-Young Step Down}

\item{marginError}{the margin of error for MDES estimation based on targeted power value}

\item{p}{the proportion of samples that are assigned to the treatment}

\item{alpha}{the family wise error rate (FWER)}

\item{numCovar.1}{number of Level 1 baseline covariates (not including block dummies)}

\item{numCovar.2}{number of Level 2 baseline covariates (set to 0 for this design)}

\item{R2.1}{a vector of length M corresponding to R^2 for M outcomes of Level 1 (R^2 = variation in the data explained by the model)}

\item{R2.2}{a vector of length M corresponding to R^2 for M outcomes of Level 2 (R^2 = variation in the data explained by the model)}

\item{ICC}{intraclass correlation}

\item{mod.type}{"c" for constant effects, "f" for fixed effects, "r" for random effects (parameter not in use at the moment)}

\item{sigma}{correlation matrix for correlations between test statistics (parameter not in use at the moment.Default is set to 0.99)}

\item{rho}{correlation between outcomes when sigma is generated}

\item{omega}{NULL (parameter in development)}

\item{tnum}{the number of test statistics (samples) for all procedures other than Westfall-Young & number of permutations for WY. The default is set at 10,000.}

\item{snum}{the number of samples for Westfall-Young. The default is set at 1,000.}

\item{ncl}{ncl the number of clusters to use for parallel processing. The default is set at 2.}

\item{num.iter}{the number of iterations to look for the optimal sample size. The default is set at 20}

\item{updateProgress}{a call back function for our internal use in our Shiny application}
}
\value{
Sample number returns
}
\description{
Sample Function
}
