% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{search_path}
\alias{search_path}
\title{Get search path}
\usage{
search_path(x)
}
\arguments{
\item{x}{pumpresult object (sample or mdes)}
}
\value{
Dataframe of search path, in sequential order.
}
\description{
Get search path
}
