% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_functions.R
\name{setup_default_parallel_plan}
\alias{setup_default_parallel_plan}
\title{Setup parallel processing}
\usage{
setup_default_parallel_plan()
}
\description{
Set up furrr to use all but one core
}
\keyword{internal}
