% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{SS.blockedRCT.2.RAW}
\alias{SS.blockedRCT.2.RAW}
\title{Calculating Raw Sample}
\usage{
SS.blockedRCT.2.RAW(J, n.j, J0 = 10, n.j0 = 10, whichSS, MDES, power,
  p, alpha, numCovar.1, numCovar.2 = 0, R2.1, R2.2, ICC, mod.type, sigma,
  omega, two.tailed = TRUE, num.iter = 100, tol = 0.1)
}
\arguments{
\item{J}{blah blah}

\item{n.j}{blah blah}

\item{J0}{starting values for J0 to look for optimal J and n.j}

\item{n.j0}{starting values for n.j0 to look for optimal J and n.j}

\item{whichSS}{blah blah}

\item{MDES}{blah blah}

\item{power}{blah blah}

\item{p}{blah blah}

\item{alpha}{blah blah}

\item{numCovar.1}{blah blah}

\item{numCovar.2}{blah blah}

\item{R2.1}{blah blah}

\item{R2.2}{blah blah}

\item{ICC}{blah blah}

\item{mod.type}{blah blah}

\item{sigma}{blah blah}

\item{omega}{blah blah}

\item{two.tailed}{blah blah}

\item{num.iter}{blah blah}

\item{tol}{blah blah}
}
\value{
raw sample returns
}
\description{
# this is a help function for getting SS when no adjustment -
it starts with PowerUp package function mrss.bira2cl but that function seems to have a bug - only works if pass in numeric values and not if pass in objects that hold those values. Plus mrss.bira2cl only computes J, not n.j
}
