% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{SS.blockedRCT.2}
\alias{SS.blockedRCT.2}
\title{Sample Function}
\usage{
SS.blockedRCT.2(M, numFalse, typesample, J, n.j, J0, n.j0, MDES, power,
  power.definition, MTP, marginError, p, alpha, numCovar.1,
  numCovar.2 = 0, R2.1, R2.2, ICC, mod.type, sigma, omega,
  tnum = 10000, snum = 2, ncl = 2, num.iter = 20,
  updateProgress = NULL)
}
\arguments{
\item{M}{blah blah}

\item{numFalse}{blah blah}

\item{typesample}{blah blah}

\item{J}{blah blah}

\item{n.j}{blah blah}

\item{J0}{blah blah}

\item{n.j0}{blah blah}

\item{MDES}{blah blah}

\item{power}{blah blah}

\item{power.definition}{blah blah}

\item{MTP}{blah blah}

\item{marginError}{blah blah}

\item{p}{blah blah}

\item{alpha}{blah blah}

\item{numCovar.1}{blah blah}

\item{numCovar.2}{blah blah}

\item{R2.1}{blah blah}

\item{R2.2}{blah blah}

\item{ICC}{blah blah}

\item{mod.type}{blah blah}

\item{sigma}{blah blah}

\item{omega}{blah blah}

\item{tnum}{blah blah}

\item{snum}{blah blah}

\item{ncl}{blah blah}

\item{num.iter}{blah blah}

\item{updateProgress}{blah blah}
}
\value{
Sample number returns
}
\description{
Sample Function
}
