% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{strip_SEs}
\alias{strip_SEs}
\title{Remove SE and df columns from (wide) power table}
\usage{
strip_SEs(power_table)
}
\arguments{
\item{power_table}{Dataframe (power result object).}
}
\value{
Changed dataframe with all columns starting with SE or df
  dropped.
}
\description{
This is used to reduce the info on a power table before pivoting to
long format.
}
\keyword{internal}
