% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{summary.pumpresult}
\alias{summary.pumpresult}
\title{Pretty print pump result with parameters}
\usage{
\method{summary}{pumpresult}(x, ...)
}
\arguments{
\item{x}{A pumpresult object.}

\item{...}{Extra options passed to print.pumpresult}
}
\description{
Pretty print pump result with parameters
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{design}()},
\code{\link{dim.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{print.pumpresult}()},
\code{\link{pump_type}()},
\code{\link{search_path}()}
}
\concept{pumpresult}
