% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{supported_designs}
\alias{supported_designs}
\title{List all the supported designs of the `pum` package.}
\usage{
supported_designs(comment = TRUE)
}
\arguments{
\item{comment}{= TRUE prints out description of each design or method.  FALSE does not.}
}
\description{
List all supported designs, with brief descriptions.
}
