% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{t.mean.H1}
\alias{t.mean.H1}
\title{t.mean.h1 helper function}
\usage{
\method{t}{mean.H1}(MDES, J, n.j, R2.1, p)
}
\arguments{
\item{MDES}{blah blah}

\item{J}{blah blah}

\item{n.j}{blah blah}

\item{R2.1}{blah blah}

\item{p}{blah blah}
}
\value{
blah blah
}
\description{
t.mean.h1 helper function
}
