% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{transpose_power_table}
\alias{transpose_power_table}
\title{Convert power table from wide to long}
\usage{
transpose_power_table(power_table, M = NULL)
}
\arguments{
\item{power_table}{pumpresult object for a power result (not mdes or sample).
(It can also take a raw dataframe of the wide table to convert to long, as
an internal helper method.)}

\item{M}{Set if power_table is a data.frame without set number of outcomes.
Usually ignore this.}
}
\description{
Transform table returned from pump_power to a long format table or to a wide
format table.
}
\examples{
pp <- pump_power(d_m = "d3.2_m3ff2rc",
  MTP = 'HO', nbar = 50, J = 30, K = 10,
  M = 5, MDES = 0.125, Tbar = 0.5, alpha = 0.05,
  numCovar.1 = 1, numCovar.2 = 1,
  R2.1 = 0.1, R2.2 = 0.1, ICC.2 = 0.2, ICC.3 = 0.2,
  omega.2 = 0, omega.3 = 0.1, rho = 0.5)
  
transpose_power_table(pp)
}
