% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpgridresult.R
\name{update.pumpgridresult}
\alias{update.pumpgridresult}
\title{Update a pump grid call, tweaking some parameters (core
  function)}
\usage{
\method{update}{pumpgridresult}(object, ...)
}
\description{
Works on objects returned by `update_grid()`; calls
  `update_grid()`.
}
\seealso{
[update_grid()]
}
