% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{update_grid}
\alias{update_grid}
\title{Update a single pump call to a grid call}
\usage{
update_grid(x, ...)
}
\arguments{
\item{x}{Pump result object.}

\item{...}{List of parameters to expand into a grid.}
}
\value{
result of calling corresponding grid
}
\description{
Given a few lists of parameters, take a pumpresult and call a grid to explore
various versions of the initial scenario.
}
\examples{
m <- pump_mdes(d_m = "d2.1_m2fc", MTP = "HO",
  nbar = 200, J = 20, power.definition = "complete",
  M = 3, target.power = 0.5, tol = 0.02,
  Tbar = 0.50, alpha = 0.05, numCovar.1 = 5,
  R2.1 = 0.1, ICC.2 = 0.05, rho = 0,
  final.tnum = 1000 )

gd <- update_grid( m, J = c( 10, 20, 30 ) )

}
