% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{validate_d_m}
\alias{validate_d_m}
\title{Validate d_m string}
\usage{
validate_d_m(d_m)
}
\value{
Full d_m string that will be found in `pump_info()`
}
\description{
Ensure d_m is a supported pair of design and model.  
If d_m is just a design, select a default model.  
Convert PowerUp! names to our naming system as needed.
}
\keyword{internal}
