% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{as.data.frame.pumpresult}
\alias{as.data.frame.pumpresult}
\alias{pumpresult}
\title{Cast pumpresult result to data.frame}
\usage{
\method{as.data.frame}{pumpresult}(x)
}
\arguments{
\item{x}{the pumpresult object to covert}
}
\description{
Cast pumpresult result to data.frame
}
\seealso{
Other pumpresult: 
\code{\link{dim.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{print.pumpresult}()},
\code{\link{search_path}()}
}
\concept{pumpresult}
