% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{calc.df}
\alias{calc.df}
\title{Calculate the degrees of freedom for a particular design}
\usage{
calc.df(
  design,
  J,
  K,
  nbar,
  numCovar.1,
  numCovar.2,
  numCovar.3,
  validate = TRUE
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}
}
\value{
Degree of freedom for the design.
}
\description{
Given sample sizes, return the used degrees of freedom (frequently
conservative) for the design.
}
