% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{calc.J}
\alias{calc.J}
\title{This function calculates J for all implemented designs}
\usage{
calc.J(design, MT, MDES, nbar, Tbar, R2.1, R2.2, ICC.2, omega.2)
}
\arguments{
\item{design}{RCT design (see list/naming convention)}

\item{nbar}{units per block}

\item{R2.1}{a vector of length M corresponding to R^2 for Level-1 covariates for M outcomes}

\item{R2.2}{a vector of length M corresponding to R^2 for Level-2 covariates for M outcomes}

\item{ICC.2}{a vector of length M of school intraclass correlation}

\item{J}{the number of schools}

\item{K}{the number of districts}

\item{R2.3}{a vector of length M corresponding to R^2 for Level-3 covariates for M outcomes}

\item{ICC.3}{a vector of length M of district intraclass correlation}

\item{numCovar.1}{number of Level 1 baseline covariates (not including block dummies)}

\item{numCovar.2}{number of Level 2 baseline covariates}

\item{numCovar.3}{number of Level 3 baseline covariates}
}
\value{
the degree of freedom
}
\description{
This function calculates J for all implemented designs
}
