% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_wy.R
\name{comp.rawt.sd}
\alias{comp.rawt.sd}
\title{Helper function for Westfall Young Step Down}
\usage{
comp.rawt.sd(nullt, rawt, rawt.order)
}
\arguments{
\item{nullt}{a vector of permuted test statistic values under H0}

\item{rawt}{a vector of raw test statistic values under H1}

\item{rawt.order}{order of raw test statistics in descending order}
}
\value{
returns a vector of 1s and 0s with lengths of M outcomes
}
\description{
Helper function for Westfall Young Step Down
}
