% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_wy.R
\name{comp.rawt.ss}
\alias{comp.rawt.ss}
\title{Helper function for Westfall Young Single Step}
\usage{
comp.rawt.ss(nullt, rawt)
}
\arguments{
\item{abs.Zs.H0.1row}{A vector of permuted test statistics values under H0}

\item{abs.Zs.H1.1samp}{One sample of raw statistics}
}
\value{
returns a vector of 1s and 0s with length of M outcomes
}
\description{
The  function  comp.rawt.SS is  needed  to  implement  the  Westfall-Young single-step multiple
testing procedure (MTP). It operates on one row of null test statistics.
}
