% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_wy.R
\name{comp.rawt.ss}
\alias{comp.rawt.ss}
\title{Helper function for Westfall Young Single Step}
\usage{
comp.rawt.ss(nullt, rawt)
}
\arguments{
\item{nullt}{a vector of permuted test statistic values under H0}

\item{rawt}{a vector of raw test statistic values under H1}
}
\value{
returns a vector of 1s and 0s with length of M outcomes
}
\description{
The  function  comp.rawt.SS is  needed  to  implement  the  Westfall-Young single-step multiple
testing procedure (MTP). It operates on one row of null test statistics.
It compares whether any of the null values across outcomes
exceeds each raw value for each outcome
}
