% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{convert_params}
\alias{convert_params}
\title{Converts model-based parameters into DGP parameters (simulation function)}
\usage{
convert_params(model.params.list)
}
\arguments{
\item{model.params.list}{list of DGP parameters}
}
\value{
dgp.params.list
}
\description{
Converts user-provided parameters such as ICC
and omega into data-generating parameters that can 
produce simulated data,
such as variance values and covariate coefficients.
}
