% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_i1_2cfr.R
\name{df}
\alias{df}
\title{Degrees of Freedom}
\usage{
df(J, n.j, numCovar.1)
}
\arguments{
\item{J}{number of blocks}

\item{n.j}{units per block}

\item{numCovar.1}{number of Level 1 baseline covariates (not including block dummies)}
}
\value{
the degree of freedom
}
\description{
This function calculates the degree of freedom for 2 level RCT block design
}
