% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{dim.pumpresult}
\alias{dim.pumpresult}
\title{Dimension of pumpresult}
\usage{
\method{dim}{pumpresult}(x, ...)
}
\description{
Dimension of pumpresult
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{power_curve}()},
\code{\link{print.pumpresult}()},
\code{\link{search_path}()}
}
\concept{pumpresult}
