% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_power.R
\name{find_best}
\alias{find_best}
\title{Extract roots from quadratic curve based on given evaluated points}
\usage{
find_best(test.pts, gamma = 1.5, target.power)
}
\arguments{
\item{test.pts}{power evaluated at different points}

\item{target.power}{goal power}

\item{start.low}{lower bound}

\item{start.high}{upper bound}

\item{alternate}{alternate point to return if quadratic fit fails}
}
\value{
root of quadratic curve
}
\description{
Extract roots from quadratic curve based on given evaluated points
}
