% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_power.R
\name{find_best}
\alias{find_best}
\title{Determine next point to check for correct power level.}
\usage{
find_best(test.pts, target.power, gamma = 1.5)
}
\arguments{
\item{test.pts}{power evaluated at different points}

\item{target.power}{goal power}

\item{gamma}{Number > 1. The amount we can extend our search range up (gamma)
or down (1/gamma) if we want to search outside our current boundaries.  1
means do not extend (which will potentially cause search to stall, not
recommended).}
}
\value{
List of estimate of when curve reaches target.power, derivative of
  curve at that point, and parameters of the fit curve.
}
\description{
Extract roots from power curve fit to given evaluated points
}
\keyword{internal}
