% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_corr_matrix}
\alias{gen_corr_matrix}
\title{Generate correlation matrix (simulation function)}
\usage{
gen_corr_matrix(M, rho.scalar)
}
\arguments{
\item{M}{scalar; dimension of matrix.}

\item{rho.scalar}{scalar; rho value.}
}
\value{
matrix; M x M correlation matrix
with rho.scalar as diagonal.
}
\description{
Generate correlation matrix (simulation function)
}
