% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_full_data}
\alias{gen_full_data}
\title{Generate simulated multi-level data (simulation function)}
\usage{
gen_full_data(dgp.params.list)
}
\arguments{
\item{dgp.params.list}{list of data generating parameters.}
}
\value{
list; potential outcomes given control y0, treatment y1,
        covariates V.k, X.jk, C.ijk.
}
\description{
Generates simulated data for multi-level
RCTs for pump-suppored designs and models for
both unobserved and observed potential outcomes.

Takes in a list of necessary data-generating parameters.

This function is beyond the main scope of calculating power,
and is instead used for simulating data.
For more info on use, see the simulation vignette.
}
