% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_simple_assignments}
\alias{gen_simple_assignments}
\title{Generates school and district ID assignments (simulations)}
\usage{
gen_simple_assignments(J, K, nbar)
}
\arguments{
\item{J}{number of schools per district}

\item{K}{number of districts}

\item{nbar}{number of individuals per school}
}
\value{
list(S.id, D.id) of school and district 
assignments for each individual
}
\description{
Generates simple default schools and 
districts IDs for individual students for the purpose of
simulations. This assumes equal sized schools in 
equal sized districts.
}
