% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_T.x}
\alias{gen_T.x}
\title{Generate treatment assignment vector (simulation function)}
\usage{
gen_T.x(d_m, S.id, D.id, nbar, Tbar)
}
\arguments{
\item{d_m}{string; design and model.}

\item{S.id}{vector; school assignments.}

\item{D.id}{vector; district assignments.}

\item{nbar}{scalar; number of level 1 units.}

\item{Tbar}{scalar; probability of treatment assignment.}
}
\value{
vector; treatment assignments for each unit.
}
\description{
Given a RCT design and supporting information,
generates treatment assignments for each student.

This function is beyond the main scope of calculating power,
and is instead used for simulating data.
For more info on use, see the simulation vignette.
}
