% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_Yobs}
\alias{gen_Yobs}
\title{convert full potential outcomes to Yobs}
\usage{
gen_Yobs(full.data, T.x)
}
\arguments{
\item{full.data}{full dataset of potential outcomes}

\item{T.x}{N-vector of binary assignment to treat/control}
}
\value{
Yobs
}
\description{
convert full potential outcomes to Yobs
}
