% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_Yobs}
\alias{gen_Yobs}
\title{Generate observed outcomes (simulation function)}
\usage{
gen_Yobs(full.data, T.x)
}
\arguments{
\item{full.data}{data.frame; full dataset of potential outcomes.}

\item{T.x}{vector; binary assignment to treat/control.}
}
\value{
vector; observed outcomes
}
\description{
Takes in a full dataset of both observed and latent
  potential outcomes and the treatment assignment vector, and
  returns only the observed outcomes.

  This function is beyond the main scope of calculating power, and
  is instead used for simulating data. For more info on use, see
  the simulation vignette.
}
