% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_wy.R
\name{get.adjp.minp}
\alias{get.adjp.minp}
\title{Helper function for Westfall Young}
\usage{
get.adjp.minp(ind.B, rawt.order)
}
\arguments{
\item{ind.B}{FILL IN (Matrix)}

\item{rawt.order}{FILL IN}
}
\value{
returns adjp matrix
}
\description{
enforces monotonicity in p-values.
}
