% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{get_pval_tstat}
\alias{get_pval_tstat}
\title{Function: get_pval_tstat	                                       
 
extracts p-value and t statistics from a given fitted model.}
\usage{
get_pval_tstat(mod, d_m, model.params.list)
}
\arguments{
\item{mod}{model object}

\item{d_m}{design/model}

\item{model.params.list}{list of model parameters}
}
\description{
Function: get_pval_tstat	                                       
 
extracts p-value and t statistics from a given fitted model.
}
\keyword{Function:}
\keyword{a}
\keyword{and}
\keyword{extracts}
\keyword{from}
\keyword{get_pval_tstat}
\keyword{given}
\keyword{internal}
\keyword{model}
\keyword{p-value}
\keyword{statistics}
\keyword{t}
