% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{make_model}
\alias{make_model}
\title{Function: make.model}
\usage{
make_model(dat.m, d_m)
}
\arguments{
\item{dat.m}{a data frame for a single outcome}

\item{d_m}{design/model}
}
\description{
Function to generate and fit a model to a simulated dataset.
}
\keyword{internal}
