% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{makelist_samp}
\alias{makelist_samp}
\title{Convert multi-outcome data structure to list for each outcome.}
\usage{
makelist_samp(samp.obs, T.x)
}
\arguments{
\item{samp.obs}{a single iteration of observed data}

\item{T.x}{vector of treatment assignments}
}
\value{
List of dataframes.
}
\description{
Given the simulated multi-outcome structure, make a list of
complete (tidy) rectangular datasets, one for each outcome.
}
\keyword{internal}
