% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_power.R
\name{optimize_power}
\alias{optimize_power}
\title{Optimizes power to help in search for MDES or SS}
\usage{
optimize_power(
  design,
  search.type,
  MTP,
  target.power,
  power.definition,
  tol,
  start.tnum,
  start.low,
  start.high,
  MDES = NULL,
  J = NULL,
  K = 1,
  nbar = NULL,
  M = M,
  Tbar = Tbar,
  alpha,
  numCovar.1 = 0,
  numCovar.2 = 0,
  numCovar.3 = 0,
  R2.1 = 0,
  R2.2 = 0,
  R2.3 = 0,
  ICC.2 = 0,
  ICC.3 = 0,
  omega.2 = 0,
  omega.3 = 0,
  rho,
  B = NULL,
  cl = NULL,
  max.steps = 20,
  max.tnum = 2000,
  final.tnum = 4 * max.tnum,
  give.warnings = FALSE,
  grid.only = FALSE,
  grid.size = 5
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{search.type}{type of optimization search being conducted. Options: K, J, nbar, mdes}

\item{MTP}{Multiple adjustment procedure of interest. Supported options:
none, Bonferroni, BH, Holm, WY-SS, WY-SD (passed as strings).  Provide list
to automatically re-run for each procedure on the list.}

\item{target.power}{Target power to arrive at}

\item{power.definition}{must be a valid power type outputted by power
function, i.e. D1indiv, min1, etc.}

\item{tol}{tolerance for target power, defaults to 0.01 (1%).  This parameter
controls when the search is done: when estimated power (checked with
`final.tnum` iterations) is within `tol`, the search stops.}

\item{start.tnum}{number of samples for first iteration of search algorithm}

\item{start.low}{lower bound for optimization procedure}

\item{start.high}{upper bound for optimization procedure}

\item{MDES}{scalar or vector; the desired MDES values for each outcome. Please
provide a scalar, a vector of length M, or vector of values for non-zero
outcomes.}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{M}{scalar; the number of hypothesis tests (outcomes), including zero
outcomes}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar, or vector of length M; school intraclass correlation}

\item{ICC.3}{scalar, or vector length M; district intraclass correlation}

\item{omega.2}{scalar, or vector of length M; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}

\item{omega.3}{scalar, or vector of length M; ratio of variance of district-average impacts to
variance of district-level random intercepts. Default to 0 (no treatment
variation).}

\item{rho}{scalar; assumed correlation between all pairs of test statistics.}

\item{B}{scalar; the number of samples/permutations for Westfall-Young}

\item{cl}{cluster object to use for parallel processing}

\item{max.steps}{how many steps allowed before terminating}

\item{max.tnum}{maximum cumulative number of samples}

\item{final.tnum}{number of samples for final draw}

\item{give.warnings}{whether to return optimizer warnings}

\item{grid.only}{TRUE means generate a grid from start.low to start.high, but
do not do iterative search. (Useful for mapping out the power curve rather
than identifying a point of particular power).}

\item{grid.size}{Number of points to check in initial search grid.  Grid will
be spaced as a quadratic sequence (e.g., 0, 1, 4, 9, 16 for a 0-16 span).}
}
\value{
power
}
\description{
Optimizes power to help in search for MDES or SS
}
