% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{params}
\alias{params}
\title{Get parameters for pump result}
\usage{
params(x, ...)
}
\value{
List of design parameters used.
}
\description{
Get parameters for pump result
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{dim.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{print.pumpresult}()},
\code{\link{search_path}()}
}
\concept{pumpresult}
