% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{parse_design}
\alias{parse_design}
\title{Return characteristics of a given design code}
\usage{
parse_design(design)
}
\arguments{
\item{design}{String. Experimental design to parse.}
}
\value{
List of features including number of levels, level of randomization,
  etc.
}
\description{
See the pump_info method to get a list of supported designs.
}
\examples{
supported = pump_info()$Design
supported$Code[[4]]
parse_design( supported$Code[[4]] )

}
\concept{pump_info}
