% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pumpgridresult.mdes}
\alias{plot.pumpgridresult.mdes}
\title{Plot a grid pump mdes object}
\usage{
\method{plot}{pumpgridresult.mdes}(x, power.definition, var.vary, ...)
}
\arguments{
\item{x}{pumpgridresult object}

\item{power.definition}{user must choose one definition of power to plot}

\item{var.vary}{a single parameter to show varying on the plot}

\item{...}{Additional parameters}
}
\description{
Plot a grid pump mdes object
}
