% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pumpgridresult.power}
\alias{plot.pumpgridresult.power}
\title{Plot a grid pump power object}
\usage{
\method{plot}{pumpgridresult.power}(x, power.definition = NULL, var.vary, ...)
}
\arguments{
\item{x}{pumpgridresult object}

\item{power.definition}{definition of power to plot.  If NULL, plot all
definitions as a facet wrap.}

\item{var.vary}{variable to vary on X axis}

\item{...}{additional parameters}
}
\value{
a ggplot object
}
\description{
Plot a grid pump power object
}
