% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pumpgridresult.power}
\alias{plot.pumpgridresult.power}
\title{Plot a grid pump power object}
\usage{
\method{plot}{pumpgridresult.power}(x, power.definition, var.vary, ...)
}
\arguments{
\item{x}{pumpgridresult object}

\item{power.definition}{user must choose one definition of power to plot}

\item{var.vary}{a single parameter to show varying on the plot}

\item{...}{Additional parameters}
}
\description{
Plot a grid pump power object
}
