% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pumpgridresult.power}
\alias{plot.pumpgridresult.power}
\title{Plot a grid pump power object}
\usage{
\method{plot}{pumpgridresult.power}(
  x,
  power.definition = NULL,
  var.vary = NULL,
  lines = TRUE,
  include.title = FALSE,
  ...
)
}
\arguments{
\item{x}{pumpgridresult object.}

\item{power.definition}{string; definition of power to plot. If
NULL, plot all definitions as a facet wrap.}

\item{var.vary}{string; variable to vary on X axis. If NULL, and
only one thing varies, then it will default to single varying
parameter.}

\item{lines}{logical; TRUE means connect dots with lines on the
plots. FALSE means no lines.}

\item{include.title}{logical; whether to include/exclude title (if
planning a facet wrap, for example).}

\item{...}{additional parameters.}
}
\value{
Plot; a ggplot object
}
\description{
Plot a grid pump power object
}
\keyword{internal}
