% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pumpresult}
\alias{plot.pumpresult}
\title{Plot a single scenario pump object}
\usage{
\method{plot}{pumpresult}(x)
}
\arguments{
\item{x}{pumpresult object}
}
\description{
Plot a single scenario pump object
}
