% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power_curve}
\alias{plot_power_curve}
\title{Examine search path of the power search.}
\usage{
plot_power_curve(pwr, plot.points = TRUE)
}
\arguments{
\item{pwr}{Result from the pump_sample or pump_mdes}

\item{plot.points}{flag; whether to plot individual points on curve}
}
\description{
This will give a plot of power vs. mdes or sample size.
It can be useful to see how quickly power changes as a function of these design parameters.
Can also be useful to gauge where convergence went poorly.
}
