% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_power.R
\name{plot_power_curve}
\alias{plot_power_curve}
\title{Examine search path of the power search.}
\usage{
plot_power_curve(pwr, plot.points = TRUE)
}
\arguments{
\item{pwr}{Result from the pump_sample or pump_mdes}

\item{plot.points}{flag; whether to plot individual points on curve}
}
\description{
This will give two plots about how the search narrowed down into the final
estimate.  Can be useful to gauge where convergence went poorly.
}
