% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_power.R
\name{plot_power_search}
\alias{plot_power_search}
\title{Examine search path of the power search}
\usage{
plot_power_search(pwr, step_plot = FALSE)
}
\arguments{
\item{pwr}{Result from the pump_sample or pump_mdes}
}
\description{
Examine search path of the power search
}
