% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power_search}
\alias{plot_power_search}
\title{Examine search path of a power search (result function)}
\usage{
plot_power_search(pwr, fit = NULL, target.line = NULL)
}
\arguments{
\item{pwr}{pumpresult object; result from a 
pump_sample or pump_mdes call.}

\item{fit}{a fitted curve to the search.}

\item{target.line}{scalar; if non-NULL, add a reference line 
for the true power (if known, e.g., from a pump_power call).}
}
\value{
plot; a ggplot object
 (a ggpubr arrangement of 3 plots, technically) of the
 search path.
}
\description{
This will give triple-plots about 
how the search narrowed down into the
final estimate.  Can be useful to gauge where 
convergence went poorly.
}
\examples{
J <- pump_sample(d_m = "d2.1_m2fc",
   MTP = 'HO', power.definition = 'D1indiv',
   typesample = 'J', target.power = 0.6,
   nbar = 50, M = 3, MDES = 0.125,
   Tbar = 0.5, alpha = 0.05,
   numCovar.1 = 1, R2.1 = 0.1, ICC.2 = 0.05, 
   rho = 0.2, tnum = 1000)
plot_power_search(J)

}
