% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrct2_power.R
\name{power.blockedRCT.2}
\alias{power.blockedRCT.2}
\title{Block RCT2 power function}
\usage{
power.blockedRCT.2(M, MDES, Ai, J, n.j, p, alpha, numCovar.1,
  numCovar.2 = 0, R2.1, R2.2 = NULL, ICC, mod.type, sigma = 0,
  rho = 0.99, omega = NULL, tnum = 10000, snum = 1000, ncl = 2,
  updateProgress = NULL)
}
\arguments{
\item{M}{the number of hypothesis tests (outcomes)}

\item{MDES}{a vector of length M corresponding to the MDESs for the M outcomes}

\item{Ai}{a single entry vector specifying the estimated number of outcomes with a non-zero effect}

\item{J}{the number of blocks}

\item{n.j}{the harmonic mean of the number of units per block}

\item{p}{the proportion of samples that are assigned to the treatment}

\item{alpha}{the family wise error rate (FWER)}

\item{numCovar.1}{number of Level 1 baseline covariates (not including block dummies)}

\item{numCovar.2}{number of Level 2 baseline covariates (set to 0 for this design)}

\item{R2.1}{a vector of length M corresponding to R^2 for M outcomes of Level 1 (R^2 = variation in the data explained by the model)}

\item{R2.2}{a vector of length M corresponding to R^2 for M outcomes of Level 2 (R^2 = variation in the data explained by the model)}

\item{ICC}{intraclass correlation}

\item{mod.type}{"c" for constant effects, "f" for fixed effects, "r" for random effects (parameter not in use at the moment)}

\item{sigma}{correlation matrix for correlations between test statistics (parameter not in use at the moment.Default is set to 0.99)}

\item{rho}{correlation between outcomes}

\item{omega}{NULL (parameter in development)}

\item{tnum}{the number of test statistics (samples) for all procedures other than Westfall-Young & number of permutations for WY. The default is set at 10,000}

\item{snum}{the number of samples for Westfall-Young. The default is set at 1,000.}

\item{ncl}{the number of clusters to use for parallel processing. The default is set at 2.}

\item{updateProgress}{the callback function to update the progress bar (User does not have to input anything)}
}
\value{
power results across all definitions of power and MTP
}
\description{
This functions calculates power for all definitions of power (individual, d-minimal, complete) for all the different MTPs
(Bonferroni, Holms, Bejamini-Hocheberg, Westfall-Young Single Step, Westfall-Young Step Down). The function works for
2-Level Block RCT
}
