% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{power_curve}
\alias{power_curve}
\title{Obtain recheck path (to see rate of power change)}
\usage{
power_curve(x, ...)
}
\arguments{
\item{x}{A pumpresult object}

\item{Dataframe}{describing power curve.}
}
\description{
Obtain recheck path (to see rate of power change)
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{design}()},
\code{\link{dim.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{print.pumpresult}()},
\code{\link{search_path}()}
}
\concept{pumpresult}
