% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{power_curve}
\alias{power_curve}
\title{Obtain power curve over a range of parameters}
\usage{
power_curve(
  x,
  all = FALSE,
  low = NULL,
  high = NULL,
  grid.size = 5,
  tnum = 2000
)
}
\arguments{
\item{x}{A pumpresult object.}

\item{all}{If TRUE, merge in the search path from the original search.}

\item{low}{Low range for the plot x-axis.}

\item{high}{High range for the plot.}

\item{grid.size}{Number of points to calculate power.}

\item{tnum}{Number of iterations to calculate power at each grid point.}
}
\description{
This is used to see rate of power change as a function of sample size or
MDES.
}
