% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{power_curve}
\alias{power_curve}
\title{Obtain power curve over a range of parameters (result function)}
\usage{
power_curve(
  x,
  all = FALSE,
  low = NULL,
  high = NULL,
  grid.size = 5,
  tnum = 2000
)
}
\arguments{
\item{x}{a pumpresult object.}

\item{all}{logical; if TRUE, merge in the search path 
from the original search.}

\item{low}{scalar; low range for the plot x-axis.}

\item{high}{scalar; high range for the plot.}

\item{grid.size}{scalar; number of points to calculate power.}

\item{tnum}{scalar; number of iterations to calculate 
power at each grid point.}
}
\value{
data.frame of power results.
}
\description{
This is used to see rate of 
power change as a function of sample size or MDES.
}
