% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{print_context}
\alias{print_context}
\title{Print context (design, model, parameter values) of 
pumpresult or pumpgridresult}
\usage{
print_context(x, insert_results = FALSE, insert_control = FALSE, ...)
}
\arguments{
\item{x}{A pumpresult object or pumpgridresult object.}

\item{insert_results}{Include actual results in the printout.}

\item{insert_control}{Include the optimizer control parameter information.}

\item{...}{Extra arguments to pass to print.pumpresult.}
}
\value{
No return value; prints results.
}
\description{
Print out the context (design and model, with parameter values) of given pump
result or pump grid result object.  The "***" denotes varying values in the
printout.
}
