% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{print_design}
\alias{print_design}
\title{Print design of given pump result object}
\usage{
print_design(x, insert_results = FALSE, ...)
}
\arguments{
\item{x}{A pumpresult object.}

\item{insert_results}{Include actual results in the printout.}

\item{...}{Extra arguments to pass to print.pumpresult.}
}
\description{
Print design of given pump result object
}
