% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{print_search}
\alias{print_search}
\title{Print the search history of a pump result object}
\usage{
print_search(x, n = 10)
}
\arguments{
\item{x}{a pumpresult object (except for is.pumpresult, where it is a generic
object to check).}

\item{n}{Number of lines of search path to print, max.}
}
\value{
Number of steps in search.
}
\description{
For pump_mdes and pump_sample, print the search history.
}
