% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMP-package.R
\docType{package}
\name{PUMP-package}
\alias{PUMP}
\alias{PUMP-package}
\title{PUMP: Power Under Multiplicity Project}
\description{
Estimates power, minimum detectable effect size (MDES) and sample size requirements. The context is multilevel randomized experiments with multiple outcomes. The estimation takes into account the use of multiple testing procedures. Development of this package was supported by a grant from the Institute of Education Sciences (R305D170030). For a full package description, including a detailed technical appendix, see \doi{10.18637/jss.v108.i06}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MDRCNY/PUMP}
  \item \url{https://mdrcny.github.io/PUMP/}
  \item Report bugs at \url{https://github.com/MDRCNY/PUMP/issues}
}

}
\author{
\strong{Maintainer}: Luke Miratrix \email{luke_miratrix@gse.harvard.edu} (\href{https://orcid.org/0000-0002-0078-1906}{ORCID})

Authors:
\itemize{
  \item Kristen Hunter \email{kristen.hunter@unsw.edu.au} (\href{https://orcid.org/0000-0002-5678-4620}{ORCID})
  \item Zarni Htet
  \item Kristin Porter \email{kristin.porter@keporterconsulting.com}
}

Other contributors:
\itemize{
  \item MDRC [copyright holder]
  \item Institute of Education Sciences [funder]
}

}
\keyword{internal}
