% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{pump_info}
\alias{pump_info}
\title{Info about user choices}
\usage{
pump_info(comment = TRUE)
}
\arguments{
\item{comment}{TRUE/FALSE, prints out 
long description of
each design and method}
}
\description{
List user options:
designs and models (d_m), including what
parameters are relevant for each context;
multiple testing procedures;
types of power;
design and model parameters.
}
