% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{pump_info}
\alias{pump_info}
\title{Provides details about supported package features}
\usage{
pump_info(
  topic = c("all", "context", "adjustment", "power", "parameters"),
  comment = TRUE
)
}
\arguments{
\item{comment}{TRUE/FALSE, prints out 
long description of
each design and method}

\item{topic:}{What kind of info.  One of}
}
\description{
List user options:
designs and models (d_m), including what
parameters are relevant for each context;
multiple testing procedures;
types of power;
design and model parameters.
}
