% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{pump_info}
\alias{pump_info}
\title{List all the supported designs of the `pum` package.}
\usage{
pump_info(comment = TRUE)
}
\arguments{
\item{comment}{= TRUE prints out description of each design or method.  FALSE does not.}
}
\description{
List all supported designs, with brief descriptions.
}
