% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_and_models.R
\name{pump_info}
\alias{pump_info}
\title{Provides details about supported package features (core function)}
\usage{
pump_info(
  topic = c("all", "context", "adjustment", "power", "parameters"),
  comment = TRUE
)
}
\arguments{
\item{topic}{string; what kind of info. One of:
all, context, adjustment, power, parameters.}

\item{comment}{logical; prints out 
long description of
each design and method.}
}
\value{
list; a list of data frames with 
information about each topic.
}
\description{
List user options:
designs and models (d_m), including what
parameters are relevant for each context;
multiple testing procedures;
types of power;
design and model parameters.
}
\seealso{
For more detailed information about user choices,
see the manuscript \url{https://arxiv.org/abs/2112.15273},
which includes a detailed Technical Appendix
including information about the designs and models
and parameters.
}
