% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_mdes.R
\name{pump_mdes}
\alias{pump_mdes}
\title{MDES (minimum detectable effect size) function}
\usage{
pump_mdes(
  design,
  MTP = NULL,
  M,
  J,
  K = 1,
  numZero = NULL,
  target.power,
  power.definition,
  tol,
  nbar,
  Tbar,
  alpha,
  numCovar.1 = 0,
  numCovar.2 = 0,
  numCovar.3 = 0,
  R2.1 = 0,
  R2.2 = 0,
  R2.3 = 0,
  ICC.2 = 0,
  ICC.3 = 0,
  rho = NULL,
  rho.matrix = NULL,
  omega.2 = 0,
  omega.3 = 0,
  B = 1000,
  max.steps = 20,
  max.tnum = 2000,
  start.tnum = 200,
  final.tnum = 4 * max.tnum,
  cl = NULL,
  updateProgress = NULL,
  give.optimizer.warnings = FALSE,
  just.result.table = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MTP}{Multiple adjustment procedure of interest. Supported options:
none, Bonferroni, BH, Holm, WY-SS, WY-SD (passed as strings).  Provide list
to automatically re-run for each procedure on the list.}

\item{M}{scalar; the number of hypothesis tests (outcomes), including zero
outcomes}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{numZero}{Additional number of outcomes assumed to be zero. Please
provide NumZero + length(MDES) = M}

\item{target.power}{Target power to arrive at}

\item{power.definition}{must be a valid power type outputted by power
function, i.e. D1indiv, min1, etc.}

\item{tol}{tolerance for target power}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{rho}{scalar; assumed correlation between the test statistics of the
tests.}

\item{omega.2}{scalar; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}

\item{omega.3}{scalar; ratio of variance of district-average impacts to
variance of district-level random intercepts. Default to 0 (no treatment
variation).}

\item{B}{scalar; the number of samples/permutations for Westfall-Young}

\item{max.steps}{how many steps allowed before terminating}

\item{max.tnum}{maximum cumulative number of samples}

\item{final.tnum}{number of samples for final draw}

\item{cl}{cluster object to use for parallel processing}

\item{updateProgress}{the callback function to update the progress bar (User
does not have to input anything)}

\item{just.result.table}{TRUE means only return final answer, FALSE means
return search path information.}

\item{verbose}{Print out diagnostics of time, etc.}
}
\value{
mdes results
}
\description{
The minimum detectable effect size function calculates the most feasible
minimum detectable effect size for a given MTP, power and power definition.
The goal is to find the MDES value that satisfies the tolerance set in the
parameter in the power value.
}
