% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_functions.R
\name{pump_mdes_grid}
\alias{pump_mdes_grid}
\title{Run pump_mdes on combination of factors}
\usage{
pump_mdes_grid(
  design,
  MTP,
  M,
  target.power,
  power.definition,
  tol,
  nbar,
  J = 1,
  K = 1,
  numZero = NULL,
  Tbar,
  alpha,
  numCovar.1 = NULL,
  numCovar.2 = NULL,
  numCovar.3 = NULL,
  R2.1 = NULL,
  R2.2 = NULL,
  R2.3 = NULL,
  ICC.2 = NULL,
  ICC.3 = NULL,
  omega.2 = NULL,
  omega.3 = NULL,
  rho,
  verbose = FALSE,
  use_furrr = FALSE,
  drop_unique_columns = TRUE,
  ...
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MTP}{Multiple adjustment procedure of interest. Supported options:
none, Bonferroni, BH, Holm, WY-SS, WY-SD (passed as strings).  Provide list
to automatically re-run for each procedure on the list.}

\item{M}{scalar; the number of hypothesis tests (outcomes), including zero
outcomes}

\item{target.power}{Target power to arrive at}

\item{power.definition}{must be a valid power type outputted by power
function, i.e. D1indiv, min1, etc.}

\item{tol}{tolerance for target power}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{numZero}{Additional number of outcomes assumed to be zero. Please
provide NumZero + length(MDES) = M}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{omega.2}{scalar; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}

\item{omega.3}{scalar; ratio of variance of district-average impacts to
variance of district-level random intercepts. Default to 0 (no treatment
variation).}

\item{rho}{scalar; assumed correlation between the test statistics of the
tests.}

\item{verbose}{Print out diagnostics of time, etc.}
}
\description{
Run pump_mdes on combination of factors
}
\seealso{
Other pump_power_grid: 
\code{\link{pump_sample_grid}()}
}
\concept{pump_power_grid}
