% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{pump_power}
\alias{pump_power}
\title{Calculate power using PUMP method}
\usage{
pump_power(
  design,
  MTP = NULL,
  MDES,
  numZero = NULL,
  M,
  nbar,
  J = 1,
  K = 1,
  Tbar,
  alpha = 0.05,
  numCovar.1 = 0,
  numCovar.2 = 0,
  numCovar.3 = 0,
  R2.1 = 0,
  R2.2 = 0,
  R2.3 = 0,
  ICC.2 = 0,
  ICC.3 = 0,
  omega.2 = 0,
  omega.3 = 0,
  rho = NULL,
  rho.matrix = NULL,
  tnum = 10000,
  B = 3000,
  cl = NULL,
  updateProgress = NULL,
  validate.inputs = TRUE,
  long.table = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MTP}{Multiple adjustment procedure of interest. Supported options:
none, Bonferroni, BH, Holm, WY-SS, WY-SD (passed as strings).  Provide list
to automatically re-run for each procedure on the list.}

\item{MDES}{scalar or vector:  t he MDES values for each outcome. Please
provide a scalar, a vector of length M, or vector of values for non-zero
outcomes.}

\item{numZero}{Additional number of outcomes assumed to be zero. Please
provide NumZero + length(MDES) = M}

\item{M}{scalar; the number of hypothesis tests (outcomes), including zero
outcomes}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{omega.2}{scalar; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}

\item{omega.3}{scalar; ratio of variance of district-average impacts to
variance of district-level random intercepts. Default to 0 (no treatment
variation).}

\item{rho}{scalar; assumed correlation between the test statistics of the
tests.}

\item{tnum}{scalar; the number of test statistics (samples)}

\item{B}{scalar; the number of samples/permutations for Westfall-Young}

\item{cl}{cluster object to use for parallel processing}

\item{updateProgress}{the callback function to update the progress bar (User
does not have to input anything)}

\item{long.table}{TRUE for table with power as rows, correction as columns,
and with more verbose names.  See `transpose_power_table`.}

\item{verbose}{Print out diagnostics of time, etc.}
}
\value{
power results for MTP and unadjusted across all definitions of power
}
\description{
This functions calculates power for all definitions of power (individual,
d-minimal, complete) for all the different MTPs (none, Bonferroni, Holms,
Bejamini-Hocheberg, Westfall-Young Single Step, Westfall-Young Step Down).
}
