% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_functions.R
\name{pump_power_grid}
\alias{pump_power_grid}
\title{Run pump_power on combination of factors}
\usage{
pump_power_grid(
  design,
  MTP,
  MDES,
  M,
  nbar,
  J = 1,
  K = 1,
  numZero = NULL,
  Tbar,
  alpha,
  numCovar.1 = NULL,
  numCovar.2 = NULL,
  numCovar.3 = NULL,
  R2.1 = NULL,
  R2.2 = NULL,
  R2.3 = NULL,
  ICC.2 = NULL,
  ICC.3 = NULL,
  omega.2 = NULL,
  omega.3 = NULL,
  rho,
  verbose = FALSE,
  use_furrr = FALSE,
  drop_unique_columns = TRUE,
  ...
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MTP}{Multiple adjustment procedure of interest. Supported options:
none, Bonferroni, BH, Holm, WY-SS, WY-SD (passed as strings).  Provide list
to automatically re-run for each procedure on the list.}

\item{MDES}{This is *not* a list of MDES for each outcome, but rather a list
of MDES to explore. Each value will be assumed held constant across all M
outcomes.}

\item{M}{scalar; the number of hypothesis tests (outcomes), including zero
outcomes}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{numZero}{Additional number of outcomes assumed to be zero. Please
provide NumZero + length(MDES) = M}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{omega.2}{scalar; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}

\item{omega.3}{scalar; ratio of variance of district-average impacts to
variance of district-level random intercepts. Default to 0 (no treatment
variation).}

\item{rho}{scalar; assumed correlation between the test statistics of the
tests.}

\item{verbose}{TRUE means print out some text as calls processed.  FALSE do
not.}

\item{drop_unique_columns}{Drop all parameter colunms that did not vary
across the grid.}
}
\description{
This extenstion of `pump_power()` will take lists of parameter values and run
`pump_power()` on all combinations of these values.
}
\details{
It can only assume the same MDES value for all outcomes due to this.  (I.e.,
a vector of MDES values will be interpreted as a sequence of calls to
pump_power, one for each MDES value given).

Each parameter in the parameter list can be a list, not scalar.  It will
cross all combinations of the list.

These calls use furrr's future_pmap package to allow for parallel
computation.  You can use the `setup_default_parallel_plan()` method or your
own.  If you do nothing, it will default to single session.
}
