% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_sample.R
\name{pump_sample}
\alias{pump_sample}
\title{Calculate sample size}
\usage{
pump_sample(
  design,
  MTP,
  typesample,
  MDES,
  M,
  nbar = NULL,
  J = NULL,
  K = NULL,
  target.power,
  power.definition,
  alpha,
  two.tailed = TRUE,
  Tbar,
  numCovar.1 = 0,
  numCovar.2 = 0,
  numCovar.3 = 0,
  R2.1 = 0,
  R2.2 = 0,
  R2.3 = 0,
  ICC.2 = 0,
  ICC.3 = 0,
  rho,
  omega.2 = 0,
  omega.3 = 0,
  tnum = 10000,
  B = 1000,
  max.steps = 20,
  max.cum.tnum = 5000,
  start.tnum = 200,
  final.tnum = 10000,
  cl = NULL,
  updateProgress = NULL,
  max_sample_size = 10000,
  tol = 0.01
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MTP}{a single multiple adjustment procedure of interest. Supported
options: Bonferroni, BH, Holm, WY-SS, WY-SD}

\item{typesample}{type of sample size to calculate: "J", "K", or "nbar".}

\item{MDES}{scalar, or vector of length M; the MDES values for each outcome.}

\item{M}{scalar; the number of hypothesis tests (outcomes)}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{target.power}{target power to arrive at}

\item{power.definition}{must be a valid power type output by power function,
i.e. D1indiv, min1, etc.}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{two.tailed}{whether to calculate two-tailed or one-tailed power}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar; school intraclass correlation}

\item{ICC.3}{scalar; district intraclass correlation}

\item{rho}{scalar; correlation between outcomes}

\item{omega.2}{scalar; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}

\item{omega.3}{scalar; ratio of variance of district-average impacts to
variance of district-level random intercepts. Default to 0 (no treatment
variation).}

\item{tnum}{scalar; the number of test statistics (samples)}

\item{B}{scalar; the number of samples/permutations for Westfall-Young}

\item{max.steps}{how many steps allowed before terminating}

\item{max.cum.tnum}{maximum cumulative number of samples}

\item{final.tnum}{number of samples for final draw}

\item{cl}{cluster object to use for parallel processing}

\item{updateProgress}{the callback function to update the progress bar (User
does not have to input anything)}

\item{max_sample_size}{If the initial bounds fail to capture a range of
working values (e.g., for non-achievable power on one end of the search and
not the other), default to this sample size in the search.}

\item{tol}{tolerance of how close power should be to the target power.}
}
\value{
sample size results
}
\description{
Note: These currently only work if MDES is the same for all outcomes.
}
