% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_functions.R
\name{pump_sample_grid}
\alias{pump_sample_grid}
\title{Run pump_sample on combination of factors}
\usage{
pump_sample_grid(
  design,
  MTP,
  M,
  target.power,
  power.definition,
  tol,
  MDES = NULL,
  typesample,
  nbar = NULL,
  J = NULL,
  K = NULL,
  Tbar,
  alpha,
  numCovar.1 = NULL,
  numCovar.2 = NULL,
  numCovar.3 = NULL,
  R2.1 = NULL,
  R2.2 = NULL,
  R2.3 = NULL,
  ICC.2 = NULL,
  ICC.3 = NULL,
  omega.2 = NULL,
  omega.3 = NULL,
  rho,
  verbose = FALSE,
  use_furrr = FALSE,
  drop_unique_columns = TRUE,
  ...
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MTP}{Multiple adjustment procedure of interest. Supported options:
none, Bonferroni, BH, Holm, WY-SS, WY-SD (passed as strings).  Provide list
to automatically re-run for each procedure on the list.}

\item{M}{scalar; the number of hypothesis tests (outcomes), including zero
outcomes}

\item{target.power}{Target power to arrive at}

\item{power.definition}{must be a valid power type outputted by power
function, i.e. D1indiv, min1, etc.}

\item{tol}{tolerance for target power, defaults to 0.01 (1%).  This parameter
controls when the search is done: when estimated power (checked with
`final.tnum` iterations) is within `tol`, the search stops.}

\item{MDES}{scalar; the MDES value for all outcomes.}

\item{typesample}{type of sample size to calculate: "nbar", "J", or "K".}

\item{nbar}{scalar; the harmonic mean of the number of units per school}

\item{J}{scalar; the number of schools}

\item{K}{scalar; the number of districts}

\item{Tbar}{scalar; the proportion of samples that are assigned to the
treatment}

\item{alpha}{scalar; the family wise error rate (FWER)}

\item{numCovar.1}{scalar; number of Level 1 (individual) covariates (not
including block dummies)}

\item{numCovar.2}{scalar; number of Level 2 (school) covariates}

\item{numCovar.3}{scalar; number of Level 3 (district) covariates}

\item{R2.1}{scalar, or vector of length M; percent of variation explained by
Level 1 covariates for each outcome. Defaults to 0.}

\item{R2.2}{scalar, or vector of length M; percent of variation explained by
Level 2 covariates for each outcome. Defaults to 0.}

\item{R2.3}{scalar, or vector of length M; percent of variation explained by
Level 3 covariates for each outcome. Defaults to 0.}

\item{ICC.2}{scalar, or vector of length M; school intraclass correlation}

\item{ICC.3}{scalar, or vector length M; district intraclass correlation}

\item{omega.2}{scalar, or vector of length M; ratio of variance of school-average impacts to
variance of school-level random intercepts.  Default to 0 (no treatment
variation).}

\item{omega.3}{scalar, or vector of length M; ratio of variance of district-average impacts to
variance of district-level random intercepts. Default to 0 (no treatment
variation).}

\item{rho}{scalar; assumed correlation between all pairs of test statistics.}

\item{verbose}{Print out diagnostics of time, etc.}

\item{use_furrr}{Use parallel processing furrr package to fit grid.}

\item{drop_unique_columns}{Drop all parameter colunms that did not vary
across the grid.}

\item{...}{Extra arguments passed to the underlying pump_power, pump_sample,
or pump_mdes functions.}
}
\description{
Run pump_sample on combination of factors
}
\seealso{
Other grid functions: 
\code{\link{pump_mdes_grid}()},
\code{\link{pump_power_grid}()}
}
\concept{grid functions}
