% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{pump_sample_raw}
\alias{pump_sample_raw}
\title{Calculating Sample for Raw (Unadjusted)}
\usage{
pump_sample_raw(
  design,
  MTP,
  typesample,
  MDES,
  M,
  J = NULL,
  K = 1,
  J0 = 10,
  K0 = 4,
  nbar0 = 10,
  target.power,
  nbar = NULL,
  Tbar,
  alpha,
  two.tailed = TRUE,
  numCovar.1 = 0,
  numCovar.2 = 0,
  numCovar.3 = 0,
  R2.1,
  R2.2 = NULL,
  R2.3 = NULL,
  ICC.2 = NULL,
  ICC.3 = NULL,
  rho = NULL,
  omega.2 = NULL,
  omega.3 = NULL,
  tol = 0.1,
  max.steps = 100
)
}
\arguments{
\item{design}{RCT design (see list/naming convention)}

\item{MTP}{multiple adjustment procedures of interest such as Bonferroni, BH, Holms, WY_SS & WY_SD
(we expect inputs in  such order)}

\item{typesample}{the type of the number of sample we would like to estimate: either block J or nbar}

\item{M}{the number of hypothesis tests (outcomes)}

\item{J}{the number of schools}

\item{K}{the number of districts}

\item{J0}{starting values for J}

\item{K0}{starting values for K}

\item{nbar0}{starting values for nbar0 to look for optimal J and nbar}

\item{target.power}{required statistical power for the experiment}

\item{nbar}{the harmonic mean of the number of units per block}

\item{Tbar}{the proportion of samples that are assigned to the treatment}

\item{alpha}{the family wise error rate (FWER)}

\item{numCovar.1}{number of Level 1 baseline covariates (not including block dummies)}

\item{numCovar.2}{number of Level 2 baseline covariates (set to 0 for this design)}

\item{numCovar.3}{number of Level 3 baseline covariates (set to 0 for this design)}

\item{R2.1}{a vector of length M corresponding to R^2 for M outcomes of Level 1 (R^2 = variation in the data explained by the model)}

\item{R2.2}{a vector of length M corresponding to R^2 for M outcomes of Level 2 (R^2 = variation in the data explained by the model)}

\item{ICC.2}{school intraclass correlation}

\item{ICC.3}{district intraclass correlation}

\item{rho}{correlation between outcomes}

\item{omega.2}{ratio of school effect size variability to random effects variability}

\item{omega.3}{ratio of district effect size variability to random effects variability}

\item{tol}{the tolerance for MDES estimation based on targeted power value}

\item{power.definition}{definition of statistical power from individual, d-minimal to complete power}

\item{tnum}{the number of test statistics (samples) for all procedures other than Westfall-Young & number of permutations for WY. The default is set at 10,000}

\item{B}{the number of samples for Westfall-Young. The default is set at 1,000.}

\item{cl}{clusters object to use for parallel processing.}

\item{updateProgress}{the callback function to update the progress bar (User does not have to input anything)}
}
\value{
raw sample returns
}
\description{
This is a Helper function for getting Sample Size when no adjustments has been made to the test statistics.
The function starts with PowerUp package function mrss.bira2cl but that function seems to have a bug -
Only works if we pass in numeric values and not if we pass in objects that hold those values.
Additionally, mrss.bira2cl only computes J, not nbar.
}
