% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpgridresult.R
\name{pumpgridresult}
\alias{pumpgridresult}
\alias{is.pumpgridresult}
\alias{summary.pumpgridresult}
\title{pumpresult object for results of grid power calculations}
\usage{
is.pumpgridresult(x)

\method{summary}{pumpgridresult}(object, ...)
}
\arguments{
\item{x}{a pumpgridresult object (except for is.pumpresult, where it is a generic
object to check).}

\item{object}{Object to summarize.}

\item{...}{Extra options passed to print.pumpresult}
}
\value{
is.pumpgridresult: TRUE if object is a pumpgridresult object.
}
\description{
The pumpgridresult object is an S3 class that holds the results from
`pump_power_grid()`, `pump_sample_grid()`, and `pump_mdes_grid()`.

It has several methods that pull different information from this object, and
some printing methods for getting nicely formatted results.
}
