% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpgridresult.R
\name{pumpgridresult}
\alias{pumpgridresult}
\alias{is.pumpgridresult}
\alias{print.pumpgridresult}
\alias{summary.pumpgridresult}
\title{Result object for results of grid power calculations}
\usage{
is.pumpgridresult(x)

\method{print}{pumpgridresult}(x, header = TRUE, include_SE = FALSE, ...)

\method{summary}{pumpgridresult}(object, include_SE = FALSE, ...)
}
\arguments{
\item{x}{a pumpgridresult object 
(except for is.pumpgridresult, where it is a generic object to check).}

\item{header}{logical; FALSE means skip some 
header info on the result, just print
the data.frame of actual results.}

\item{include_SE}{logical; TRUE means include standard errors and df.}

\item{...}{extra options passed to print.pumpgridresult}

\item{object}{object to summarize.}
}
\value{
is.pumpgridresult: TRUE if object is a pumpgridresult object.

print: No return value; prints results.

summary: No return value; prints results.
}
\description{
The pumpgridresult object is an S3 class that holds the results from
`pump_power_grid()`, `pump_sample_grid()`, and `pump_mdes_grid()`.

It has several methods that pull different information from this object, and
some printing methods for getting nicely formatted results.
}
