% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{pumpresult}
\alias{pumpresult}
\alias{params}
\alias{design}
\alias{search_path}
\alias{pump_type}
\alias{is.pumpresult}
\alias{dim.pumpresult}
\alias{summary.pumpresult}
\alias{print.pumpresult}
\alias{as.data.frame.pumpresult}
\title{pumpresult object for results of power calculations}
\usage{
params(x, ...)

design(x, ...)

search_path(x, ...)

pump_type(x)

is.pumpresult(x)

\method{dim}{pumpresult}(x, ...)

\method{summary}{pumpresult}(object, ...)

\method{print}{pumpresult}(x, n = 10, header = TRUE, search = header, ...)

\method{as.data.frame}{pumpresult}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a pumpresult object (except for is.pumpresult, where it is a generic
object to check).}

\item{...}{additional arguments to be passed to the as.data.frame.list methods.}

\item{object}{Object to summarize.}

\item{n}{Number of lines of search path to print, max.}

\item{header}{FALSE means skip some header info on the result, just print
the data.frame of actual results.}

\item{search}{FALSE means don't print the search path for a result for
mdes or sample.}

\item{row.names}{NULL or a character vector giving the row names for the data frame.}

\item{optional}{logical. If TRUE, setting row names and converting column names is optional.}
}
\value{
params: List of design parameters used.

design: design used (as string)

search_path: Dataframe describing search path, if it was saved in the pumpresult object.

pump_type: power, mdes, or sample, as a string.

is.pumpresult: TRUE if object is a pumpresult object.

dim: Dimension of pumpresult (as matrix)

as.data.frame: pumpresult object as a clean dataframe (no more attributes from
  pumpresult).
}
\description{
The pumpresult object is an S3 class that holds the results from
`pump_power()`, `pump_sample()`, and `pump_mdes()`.

It has several methods that pull different information from this object, and
some printing methods for getting nicely formatted results.

Pump result objects are also data.frames, so they can be easily manipulated
and combined.  The return values from the `grid` functions will just return
data frames in general.
}
\seealso{
updateå

update_grid
}
