% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_functions.R
\name{run_grid}
\alias{run_grid}
\title{Run grid across any of the core pump functions}
\usage{
run_grid(
  args,
  pum_function,
  verbose = FALSE,
  drop.unique.columns,
  ...,
  use.furrr = FALSE
)
}
\arguments{
\item{args}{list of scenario arguments}

\item{pum_function}{pump_mdes, pump_sample, pump_power}

\item{verbose}{print out detailed diagnostics}

\item{drop.unique.columns}{logical; drop all parameter columns 
that did not vary across the grid.}

\item{...}{Extra arguments passed to the underlying pump_power, pump_sample,
or pump_mdes functions.}

\item{use.furrr}{not currently implemented; whether to use furr package
for parallelization}
}
\description{
Run grid across any of the core pump functions
}
\keyword{internal}
