% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{search_path}
\alias{search_path}
\title{Obtain search path of pump_mdes or pump_sample call}
\usage{
search_path(x, ...)
}
\arguments{
\item{x}{A pumpresult object}

\item{Dataframe}{describing search path, if it was saved in the pumpresult object.}
}
\description{
Obtain search path of pump_mdes or pump_sample call
}
\seealso{
Other pumpresult: 
\code{\link{as.data.frame.pumpresult}()},
\code{\link{dim.pumpresult}()},
\code{\link{is.pumpresult}()},
\code{\link{params}()},
\code{\link{power_curve}()},
\code{\link{print.pumpresult}()}
}
\concept{pumpresult}
