% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{transpose_power_table}
\alias{transpose_power_table}
\title{Convert power table from wide to long}
\usage{
transpose_power_table(power_table)
}
\arguments{
\item{power_table}{pumpresult object for a power result (not mdes or sample).}
}
\description{
Transform table returned from pump_power to a long format table.
}
