% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{transpose_power_table}
\alias{transpose_power_table}
\title{Convert power table from wide to long (result function)}
\usage{
transpose_power_table(power_table, M = NULL)
}
\arguments{
\item{power_table}{pumpresult object for a power result 
(not mdes or sample). (It can also take a raw dataframe of the wide table 
to convert to long, as an internal helper method.)}

\item{M}{scalar; set if power_table is a data.frame 
without set number of outcomes. Usually ignore this.}
}
\value{
data.frame of power results in long format.
}
\description{
Transform table returned from pump_power 
to a long format table or to a wide format table.
}
