% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{update.pumpresult}
\alias{update.pumpresult}
\title{Update a pump call, tweaking some parameters}
\usage{
\method{update}{pumpresult}(x, type = NULL, ...)
}
\arguments{
\item{x}{Pump result object.}
}
\value{
New call using parameters of old object.
}
\description{
Update a pump call, tweaking some parameters
}
