% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{update.pumpresult}
\alias{update.pumpresult}
\title{Update a pump call, tweaking some parameters}
\usage{
\method{update}{pumpresult}(object, ...)
}
\arguments{
\item{object}{Pump result object.}

\item{...}{Parameters as specified in `pump_power`, `pump_mdes`, and
`pump_sample` that should be overwritten.}
}
\value{
Results of a new call using parameters of old object with newly
  specified parameters replaced.
}
\description{
One of the optional parameters can be a `type = something` argument, where
the "something" is either "power", "sample", or "mdes", if the call should be
shifted to a different pump call (pump_power, pump_sample, or pump_mdes,
respectively).
}
