% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{update_grid}
\alias{update_grid}
\title{Update a single pump call to a grid call}
\usage{
update_grid(x, ...)
}
\arguments{
\item{x}{Pump result object.}

\item{...}{List of parameters to expand into a grid.}
}
\value{
result of calling corresponding grid
}
\description{
Take a pumpresult and provide lists 
of parameters to explore various versions 
of the initial scenario.
}
\examples{
pp <- pump_power(d_m = "d2.1_m2fc", MTP = "HO",
  nbar = 200, J = 20, MDES = 0.2, M = 3,
  Tbar = 0.50, alpha = 0.05, numCovar.1 = 5,
  R2.1 = 0.1, ICC.2 = 0.05, rho = 0)

gd <- update_grid( pp, J = c( 10, 20, 30 ) )

}
