% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpresult.R
\name{update_grid}
\alias{update_grid}
\title{Update a single pump call to a grid call}
\usage{
update_grid(x, ...)
}
\arguments{
\item{x}{Pump result object.}

\item{...}{List of parameters to expand into a grid.}
}
\value{
result of calling corresponding grid
}
\description{
Given a few lists of parameters, take a pumpresult and call a grid to explore
various versions of the initial scenario.
}
