% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pump_power.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validates user inputs}
\usage{
validate_inputs(
  design,
  MTP,
  params.list,
  mdes_call = FALSE,
  single_MDES = FALSE
)
}
\arguments{
\item{design}{a single RCT design (see list/naming convention)}

\item{MTP}{a single multiple adjustment procedure of interest.}

\item{params.list}{a list of parameters input by a user}
}
\value{
params.list
}
\description{
This functions takes in a list of user inputs. Depending on the inputs,
it produces errors or warnings, and at times modifies inputs if necessary.
}
